Remedy (Remedy v0.6.9)

This is the documentation for the Remedy library.

Configuration

config :remedy
  token: System.get_env("REMEDY_TOKEN")
  intents: 12,
  shards: :auto,

Configuration can be provided as starting arguments or inside your config.exs file. The configurable terms are as follows.

:token (required) Your bots token, available from your Application Dashboard

:shards Number of shards to use.

:intents See Remedy.Gateway.Intents for more information

:min_workers The minimum number of HTTP2 connections to keep open. Defaults to 1, cannot be zero.

:max_workers The maximum number of HTTP2 connections to keep open. Defaults to 10, cannot be zero.

Environment Variables

It is recommended to use environment variables to store sensitive configuration. So that they are not exposed if you should upload your project to GitHub etc. If a token is uploded to a public repo, Discord will notice and invalidate it.

Connection Limits

Some hosting platforms will limit the number of connections you are allowed to keep active at any one time.

Each shard and HTTP2 worker will count for an individual connection. In addition to any other connections required for your application. Phoenix will establish 10 connections to the database by default. If you are using Heroku free tier (20 connection limit) you could run out of connections for users visiting your web application. In those cases you should consider limiting the number of shards and HTTP2 workers.

We recommend using Gigalixir or Fly.io for your hosting platform. The gigalixir free tier has unlimited connections and will allow your bot to run automatically configured shards without issues.

FFmpeg

Functions

discord_api_version()

Specs

discord_api_version() :: integer()

Retreive the configured API version

intents()

Specs

intents() :: integer()

Retreive the configured gateway intents.

max_workers()

Specs

max_workers() :: integer()

Retreive the configured max HTTP worker count.

min_workers()

Specs

min_workers() :: integer()

Retreive the configured minimum HTTP worker count.

shards()

Specs

shards() :: integer()

Retreive the configured shard count

system_architecture()

Specs

system_architecture() :: binary()

Retreive the current system architecture.

token()

Specs

token() :: any()

Retreive the bot token.