Remedy.API (Remedy v0.6.8)

Standard interface for the Discord API.

The majority of the functions within this module are pulled directly from the Discord API. Some custom implementations are included.

Some functions are useless in the scope of a bot and are intentionally omitted.

Ratelimits

Discord imposes rate limits in various capacities. The functions in this module will respect those rate limits where possible. In certain circumstances, discord will impose hidden rate limits to avoid abuse, which you can still hit.

Bang!

While intentionally undocumented to reduce clutter, all functions can be banged to return or raise.

Return Values

Return values have a number of forms due to the wide range of returns from Discord. Generally they take one of the following forms.

:ok Returned when discord indicates an empty 204 response. Which indicates a request executed successfully. Usually returns for operations such as deleting a channel.

{:ok, body | [body]} Returned when the request has created or changed one or more objects.

{:error, {403, 10004, "Unknown Guild"}} Returned when a request fails.

Types

code()

Specs

code() :: integer()

error()

Specs

error() :: any()

limit()

Specs

limit() :: integer()

locator()

Specs

locator() :: any()

opts()

Specs

opts() :: keyword() | nil

params()

Specs

params() :: keyword() | nil

reason()

Specs

reason() :: String.t() | nil

snowflake()

Specs

snowflake() :: Sunbake.Snowflake.t()

token()

Specs

token() :: String.t()

Functions

add_guild_member(guild_id, user_id, opts)

Puts a user in a guild.

Permissions

CREATE_INSTANT_INVITE

MANAGE_NICKNAMES*

MANAGE_ROLES*

MUTE_MEMBERS*

DEAFEN_MEMBERS*

Events

:GUILD_MEMBER_ADD

Options

:access_token (string) - the user's oauth2 access token

:nick (string) - value to set users nickname to

:roles (list of t:Remedy.Schema.Guild.Role.id/0) - array of role ids the member is assigned

:mute (boolean) - if the user is muted

:deaf (boolean) - if the user is deafened

:access_token is always required.

Examples

iex> Remedy.API.add_guild_member(
...> 41771983423143937,
...> 18374719829378473,
...> access_token: "6qrZcUqja7812RVdnEKjpzOL4CvHBFG",
...> nick: "remedy",
...> roles: [431849301, 913809431])

add_guild_member_role(guild_id, user_id, role_id)

Adds a role to a member.

Role to add is specified by role_id. User to add role to is specified by guild_id and user_id. An optional reason can be given for the audit log.

add_thread_member(channel_id, user_id)
(since 0.6.0)

Specs

add_thread_member(snowflake(), snowflake()) :: {:error, any()} | :ok

Adds another member to a thread.

Requires the ability to send messages in the thread. Also requires the thread is not archived.

Events

:THREAD_MEMBERS_UPDATE

Examples

iex> Remedy.API.add_thread_member(channel_id, user_id)
:ok

begin_guild_prune(guild_id)

Begins a guild prune to prune members within days.

An optional reason can be provided for the guild audit log.

This endpoint requires the KICK_MEMBERS permission. It fires multiple Remedy.Consumer.guild_member_remove/0 events.

If successful, returns {:ok, %{pruned: pruned}}. Otherwise, returns a {:error, reason}.

Examples

iex> Remedy.API.begin_guild_prune(81384788765712384, 1)
{:ok, %{pruned: 0}}

create_channel(guild_id, opts)

Creates a channel for a guild.

Permissions

MANAGE_CHANNELS

Events

:CHANNEL_CREATE

Options

:name - :string, min: 2, max: 100

:type - Remedy.Schema.Channel.t/0

:topic - :integer, min: 8, max: 256

:bitrate - :integer, min: 8, max: 256

:user_limit - :integer, min: 1, max: 99, unlimited: 0

:permission_overwrites - [t:Remedy.Schema.PermissionOverwrite.t/0]

:parent_id - t:Snowflake.t/0 - Category to place the channel under.

:nsfw - :boolean

Examples

iex> Remedy.API.create_channel(81384788765712384, name: "elixir-remedy", topic: "steve's domain")
{:ok, %Remedy.Schema.Channel{guild_id: 81384788765712384}}

create_channel_invite(channel_id, opts)

Creates an invite for a guild channel.

An optional reason can be provided for the audit log.

This endpoint requires the CREATE_INSTANT_INVITE permission.

If successful, returns {:ok, invite}. Otherwise, returns a {:error, reason}.

Options

:max_age, :integer, default: 86400 - duration of invite in seconds before expiry, or 0 for never.

:max_uses, :integer, default: 0 - max number of uses or 0 for unlimited.

:temporary, :boolean, default: false - Whether the invite should grant temporary membership.

:unique, :boolean, default: false - used when creating unique one time use invites.

Examples

iex> Remedy.API.create_channel_invite(41771983423143933)
{:ok, Remedy.Schema.Invite{}}

iex> Remedy.API.create_channel_invite(41771983423143933, max_uses: 20)
{:ok, %Remedy.Schema.Invite{}}

create_dm(user_id)

Create a new DM channel with a user.

Examples

iex> Remedy.API.create_dm(150061853001777154)
{:ok, %Remedy.Schema.Channel{}}

create_emoji(guild_id, opts)
(since 0.6.0)

Specs

create_emoji(snowflake(), opts()) :: {:ok, Emoji.t()} | {:error, reason()}

Creates a new emoji for the given guild.

Permissions

:MANAGE_EMOJIS

Events

:EMOJIS_UPDATE

An optional reason can be provided for the audit log.

If successful, returns {:ok, emoji}. Otherwise, returns {:error, reason}.

Options

:name

:image

:roles

:reason

:name and :image are always required.

Examples

iex> Remedy.API.create_emoji(43189401384091,
...> name: "remedy", image: "data:image/png;base64,YXl5IGJieSB1IGx1a2luIDQgc3VtIGZ1az8=", roles: [])

create_followup(interaction_token)

Create a followup message for an interaction.

Delegates to execute_webhook/3, see the function for more details.

create_global_command()

Create a new global application command.

The new command will be available on all guilds in around an hour. If you want to test commands, use create_guild_command/2 instead, as commands will become available instantly there. If an existing command with the same name exists, it will be overwritten.

Parameters

application_id: Application ID for which to create the command. If not given, this will be fetched from Me.

command: Command configuration, see the linked API documentation for reference.

Return value

The created command. See the official reference: https://discord.com/developers/docs/interactions/slash-commands#create-global-application-command

Example

iex>  Remedy.API.create_command(%{name: "edit", description: "ed, man! man, ed", options: []})
{:ok, %Remedy.Schema.Command{}}

create_guild(opts)
(since 0.6.0)

Specs

create_guild(opts()) :: {:error, reason()} | {:ok, Remedy.Schema.Guild.t()}

Create a guild.

This endpoint can be used only by bots in less than 10 guilds.

Events

:GUILD_CREATE

Options

:name - :string, min: 2, max: 100

:icon - :image_data

:verification_level - :integer

0 - :none - unrestricted

1 - :low - must have verified email on account

2 - :medium - must be registered on Discord for longer than 5 minutes

3 - :high - must be a member of the server for longer than 10 minutes

4 - :very_high - must have a verified phone number

:default_message_notifications - :integer

:explicit_content_filter - :integer

0 - :disabled

1 - :members_without_roles

2 - :all_members

:roles - {:array, :role}

:channels - {:array, :channel}

:afl_channel_id - :snowflake

:afk_timeout - :integer, :seconds

Examples

iex> Remedy.API.create_guild(name: "Test Server For Testing")

When using the :roles parameter, the first member of the array is used to change properties of the guild's @everyone role. If you are trying to bootstrap a guild with additional roles, keep this in mind.

When using the :roles parameter, the required id field within each role object is an integer placeholder, and will be replaced by the API upon consumption. Its purpose is to allow you to overwrite a role's permissions in a channel when also passing in channels with the channels array.

When using the :channels parameter, the :position field is ignored, and none of the default channels are created.

When using the :channels parameter, the id field within each channel object may be set to an integer placeholder, and will be replaced by the API upon consumption. Its purpose is to allow you to create :GUILD_CATEGORY channels by setting the :parent_id field on any children to the category's id field. Category channels must be listed before any children.

create_guild_ban(guild_id, user_id)

Bans a user from a guild.

User to delete is specified by guild_id and user_id. An optional reason can be specified for the audit log.

create_guild_command(guild_id)
(since 0.6.0)

create_guild_from_template(template_code)
(since 0.6.0)

create_guild_role(guild_id)

Creates a guild role.

An optional reason for the audit log can be provided via reason.

This endpoint requires the MANAGE_ROLES permission. It fires a Remedy.Consumer.guild_role_create/0 event.

If successful, returns {:ok, role}. Otherwise, returns a {:error, reason}.

Options

:name (string) - name of the role (default: "new role")

:permissions (integer) - bitwise of the enabled/disabled permissions (default: @everyone perms)

:color (integer) - RGB color value (default: 0)

:hoist (boolean) - whether the role should be displayed separately in the sidebar (default: false)

:mentionable (boolean) - whether the role should be mentionable (default: false)

Examples

iex> Remedy.API.create_guild_role(41771983423143937, name: "remedy-club", hoist: true)
{:ok, %Remedy.Schema.Role{}}

create_guild_sticker(guild_id, sticker_id, opts)
(since 0.6.0)

Specs

create_guild_sticker(snowflake(), snowflake(), opts()) ::
  {:error, reason()} | {:ok, Remedy.Schema.Sticker.t()}

Creates a new sticker under the given guild.

Options

:name, :string, min: 2, max: 30 - name of the sticker (2-30 characters)

:description, :string, min: 2, max: 30 - description of the sticker (empty or 2-100 characters)

:tags, :string, max: 200 - autocomplete/suggestion tags for the sticker (max 200 characters)- :file, :file_contents - the sticker file to upload, must be a PNG, APNG, or Lottie JSON file, max 500 KB

Examples

iex> Remedy.API.create_guild_sticker(guild_id, channel_id, sticker_map)
{:ok, Sticker.t()}

iex> Remedy.API.create_guild_sticker(guild_id, channel_id, bad_sticker_map)
{:error, {404, 10060, "Invalid Form Body"}}

create_guild_template(guild_id)
(since 0.6.0)

create_interaction_response(interaction_id, interaction_token)

Create a response to an interaction received from the gateway.

Parameters

id: The interaction ID to which the response should be created.

token: The interaction token.

response: An InteractionResponse object. See the linked documentation.

Example

As an alternative to passing the interaction ID and token, the original Remedy.Schema.Interaction.t/0 can also be passed directly. See create_interaction_response/1.

create_message(channel_id, opts \\ [])

Specs

create_message(snowflake(), opts() | binary() | Embed.t()) ::
  {:error, any()} | {:ok, any()}

Posts a message to a guild text or DM channel.

Intents

:VIEW_CHANNEL

:SEND_MESSAGES

:SEND_MESSAGES_TTS (optional)

Events

t:Remedy.Gateway.Dispatch.message_create/0.

Options

:content (string) - the message contents (up to 2000 characters)

:tts (boolean) - true if this is a TTS message

:file (Path.t/0 | map) - the path of the file being sent, or a map with the following keys if sending a binary from memory

:name (string) - the name of the file

:body (string) - binary you wish to send

:embed (Remedy.Schema.Embed.t/0) - embedded rich content

:allowed_mentions - See "Allowed mentions" below

:message_reference (map) - See "Message references" below

Note: At least one of the following is required: :content, :file, :embed.

Allowed mentions

:all (default) - Ping everything as usual

:none - Nobody will be pinged

:everyone - Allows to ping @here and @everone

:users - Allows to ping users

:roles - Allows to ping roles

{:users, list} - Allows to ping list of users. Can contain up to 100 ids of users.

{:roles, list} - Allows to ping list of roles. Can contain up to 100 ids of roles.

Examples

iex> {:ok, message} = Remedy.API.create_message(872417560094732331, content: "**Doctest Message** âœ…")
...> message.content
"**Doctest Message** âœ…"

create_reaction(channel_id, message_id, emoji)
(since 0.6.0)

Specs

create_reaction(snowflake(), snowflake(), snowflake()) ::
  :ok | {:error, reason()}

Creates a reaction for a message.

Permissions

VIEW_CHANNEL

READ_MESSAGE_HISTORY

ADD_REACTIONS ()

Examples

iex> Remedy.API.create_reaction(123123123123, 321321321321, ...> %Remedy.Schema.Emoji{id: 43819043108, name: "foxbot"} ...> ) :ok

iex> Remedy.API.create_reaction(123123123123, 321321321321, "ðŸ˜")

create_stage()
(since 0.6.0)

create_webhook(channel_id)

Creates a webhook.

Parameters

channel_id - Id of the channel to send the message to.

args - Map with the following required keys:

name - Name of the webhook.

avatar - Base64 128x128 jpeg image for the default avatar.

reason - An optional reason for the guild audit log.

delete_all_reactions(channel_id, message_id)
(since 0.6.0)

Specs

delete_all_reactions(snowflake(), snowflake()) :: {:error, reason()} | :ok

Deletes all reactions from a message.

Permissions

VIEW_CHANNEL,

READ_MESSAGE_HISTORY

MANAGE_MESSAGES

Events

:MESSAGE_REACTION_REMOVE_ALL.

Examples

iex> Remedy.API.delete_all_reactions(893605899128676443, 912815032755191838) :ok

delete_all_reactions_for_emoji(channel_id, message_id, emoji)

Deletes all reactions of a given emoji from a message.

Permissions

MANAGE_MESSAGES

Events

:MESSAGE_REACTION_REMOVE_EMOJI

If successful, returns {:ok}. Otherwise, returns {:error, reason}.

See create_reaction/3 for similar examples.

delete_channel(channel_id)
(since 0.6.0)

Specs

delete_channel(snowflake()) :: {:error, any()} | {:ok, map()}

Deletes a channel.

An optional reason can be provided for the guild audit log.

Permissions

MANAGE_CHANNELS

Events

:CHANNEL_DELETE

Examples

iex> Remedy.API.delete_channel(421533712753360896) {:ok, %Remedy.Schema.Channel{id: 421533712753360896}}

iex> Remedy.API.delete_channel(123)

delete_channel_permission(channel_id, overwrite_id, reason \\ nil)

Specs

delete_channel_permission(snowflake(), snowflake(), reason()) ::
  {:error, reason()} | :ok

Delete a channel permission for a user or role.

Role or user overwrite to delete is specified by channel_id and overwrite_id. An optional reason can be given for the audit log.

delete_emoji(guild_id, emoji_id, opts)
(since 0.6.0)

Specs

delete_emoji(snowflake(), snowflake(), opts()) :: :ok | {:error, any()}

Deletes the given emoji.

Permissions

:MANAGE_EMOJIS

Events

:EMOJI_UPDATE

Options

:reason

Examples

iex> Remedy.API.delete_emoji(snowflake, snowflake, reason: "Because i felt like it")
:ok

delete_followup(interaction_token, message_id)
(since 0.6.0)

Specs

delete_followup(token(), snowflake()) :: :ok | {:error, reason()}

Delete a followup message for an interaction.

Parameters

:token - Interaction token.

:message_id - Followup message ID.

delete_global_command(command_id)

Delete an existing global application command.

Parameters

application_id: Application ID for which to create the command. If not given, this will be fetched from Me.

command_id: The current snowflake of the command.

delete_guild(guild_id)

Deletes a guild.

This endpoint requires that the current user is the owner of the guild.

Events

:GUILD_DELETE

Examples

iex> Remedy.API.delete_guild(618432108653707274)
{:error, {403, 50001, "Missing Access"}}

iex> Remedy.API.delete_guild(618432108653707274)
{:error, {403, 50001, "Missing Access"}}

delete_guild_command(guild_id, command_id)
(since 0.6.0)

delete_guild_integration(guild_id, integration_id)

Deletes a guild integeration.

Integration to delete is specified by guild_id and integeration_id.

delete_guild_role(guild_id, role_id)

Deletes a role from a guild.

An optional reason can be specified for the audit log.

This endpoint requires the MANAGE_ROLES permission. It fires a Remedy.Consumer.guild_role_delete/0 event.

If successful, returns {:ok}. Otherwise, returns a {:error, reason}.

Examples

iex> Remedy.API.delete_guild_role(41771983423143937, 392817238471936)

delete_guild_sticker(guild_id, sticker_id)
(since 0.6.0)

delete_guild_template(guild_id, template_code)
(since 0.6.0)

delete_invite(invite_code)

Deletes an invite by its invite_code.

This endpoint requires the MANAGE_CHANNELS permission.

If successful, returns {:ok, invite}. Otherwise, returns a {:error, reason}.

Examples

iex> Remedy.API.delete_invite("zsjUsC")
{:ok, %Remedy.Schema.Invite{}}

delete_message(channel_id, message_id)

Deletes a message.

Permissions

'VIEW_CHANNEL'

'MANAGE_MESSAGES'

Events

MESSAGE_DELETE

Examples

iex> Remedy.API.delete_message(43189401384091, 43189401384091)

delete_messages(channel_id, opts)

Deletes multiple messages from a channel.

messages is a list of Remedy.Schema.Message.id that you wish to delete. When given more than 100 messages, this function will chunk the given message list into blocks of 100 and send them off to the API. It will stop deleting on the first error that occurs. Keep in mind that deleting thousands of messages will take a pretty long time and it may be proper to just delete the channel you want to bulk delete in and recreate it.

This method can only delete messages sent within the last two weeks. Filter is an optional parameter that specifies whether messages sent over two weeks ago should be filtered out; defaults to true.

delete_original_interaction_response(interaction_token)
(since 0.6.0)

delete_own_reaction(channel_id, message_id, emoji)
(since 0.6.0)

Specs

delete_own_reaction(snowflake(), snowflake(), term()) ::
  :ok | {:error, reason()}

Deletes a reaction the bot has made for the message.

Permissions

VIEW_CHANNEL

READ_MESSAGE_HISTORY

Events

:MESSAGE_REACTION_REMOVE

Examples

iex> Remedy.API.delete_own_reaction(123, 123, 123) :ok

delete_stage(channel_id)
(since 0.6.0)

delete_user_reaction(channel_id, message_id, emoji, user_id)
(since 0.6.0)

Specs

delete_user_reaction(snowflake(), snowflake(), snowflake(), any()) ::
  {:error, any()} | {:ok, any()}

Deletes another user's reaction from a message.

Permissions

VIEW_CHANNEL

READ_MESSAGE_HISTORY

MANAGE_MESSAGES

Examples

iex>

delete_webhook(webhook_id)

Deletes a webhook.

Parameters

webhook_id - Id of webhook to delete.

reason - An optional reason for the guild audit log.

delete_webhook_message(webhook_id, webhook_token, message_id)
(since 0.6.0)

delete_webhooks_with_token(webhook_id, webhook_token)
(since 0.6.0)

execute_webhook(webhook_id, webhook_token)

Executes a webhook.

## Parameters

webhook_id - Id of the webhook to execute.

webhook_token - Token of the webhook to execute.

args - Map with the following required keys:

content - Message content.

file - File to send.

embeds - List of embeds to send.

username - Overrides the default name of the webhook.

avatar_url - Overrides the default avatar of the webhook.

tts - Whether the message should be read over text to speech.

wait - Whether to return an error or not. Defaults to false.

Only one of content, file or embeds should be supplied in the args parameter.

follow_news_channel(channel_id, webhook_channel_id)
(since 0.6.0)

get_application()
(since 0.6.8)

Specs

get_application() :: {:error, any()} | {:ok, map()}

Gets the bot's application info.

Examples

iex> Remedy.API.get_application()
{:ok, %Remedy.Schema.App{}

get_audit_log(guild_id, opts \\ [])
(since 0.6.0)

Specs

get_audit_log(snowflake() | Remedy.Schema.Guild.t(), opts()) ::
  {:error, any()} | {:ok, Remedy.Schema.AuditLog.t()}

Get the audit log for the given guild.

Permissions

-

Events

-

Options

:action_type - :integer - Audit Log Action Types

:before - Snowflake

:user_id - Snowflake

:limit - :integer, default: 50, min: 1, max: 100

Examples

iex> Remedy.API.get_audit_log(872417560094732328)
{:ok, %Remedy.Schema.AuditLog{}}

iex> Remedy.API.get_audit_log(872417560094732328, limit: 3, user_id: 883307747305725972)
{:ok, %Remedy.Schema.AuditLog{}}

iex> Remedy.API.get_audit_log(123)
{:error, {403, 10004, "Unknown Guild"}}

get_bot()
(since 0.6.0)

Specs

get_bot() :: {:error, reason()} | {:ok, Remedy.Schema.User.t()}

Gets info on the bot.

Examples

iex> Remedy.API.get_bot()
{:ok, %Remedy.Schema.User{
  id: 883307747305725972,
  avatar: "973a059282550a9ffaca42e795d8330b",
  username: "Remedy",
  ...
  }
}

get_channel(channel_id)
(since 0.6.0)

Specs

get_channel(snowflake() | Remedy.Schema.Channel.t()) ::
  {:error, any()} | {:ok, Remedy.Schema.Channel.t()}

Get a channel.

Examples

iex> Remedy.API.get_channel(872417560094732331)
{:ok, %Remedy.Schema.Channel.t{}}

get_channel_invites(channel_id)

Gets a list of invites for a channel.

This endpoint requires the 'VIEW_CHANNEL' and 'MANAGE_CHANNELS' permissions.

If successful, returns {:ok, invite}. Otherwise, returns a {:error, reason}.

Examples

iex> Remedy.API.get_channel_invites(43189401384091)
{:ok, [%Remedy.Schema.Invite{}]}

get_channels(guild_id)

Gets a list of guild channels.

Examples

iex> Remedy.API.get_channels(81384788765712384)
{:ok, [%Remedy.Schema.Channel{guild_id: 81384788765712384}]}

get_command_permissions(guild_id, command_id)
(since 0.6.0)

get_current_authorization_information()

Specs

get_current_authorization_information() :: {:error, reason()} | {:ok, map()}

get_emoji(guild_id, emoji_id)
(since 0.6.0)

Specs

get_emoji(snowflake(), snowflake()) :: {:error, reason()} | {:ok, Emoji.t()}

Gets an emoji for the given guild and emoji ids.

Permissions

:MANAGE_EMOJIS

Examples

iex> Remedy.API.get_emoji(guild_id, emoji_id) {:ok, %Emoji{}}

iex Remedy.API.get_emoji(guild_id, bad_emoji_id)

get_followup(interaction_token, message_id)
(since 0.6.0)

get_global_command(command_id)
(since 0.6.0)

get_global_commands()

Fetch all global commands.

Parameters

application_id: Application ID for which to search commands. If not given, this will be fetched from Me.

Return value

A list of ApplicationCommands on success. See the official reference: https://discord.com/developers/docs/interactions/slash-commands#applicationcommand

Example

iex> Remedy.API.get_global_commands
{:ok, [%{application_id: "455589479713865749"}]}

get_guild(guild_id)
(since 0.6.0)

Gets a guild.

Examples

iex> Remedy.API.get_guild(81384788765712384)
{:ok, %Remedy.Schema.Guild{id: 81384788765712384}}

get_guild_ban(guild_id, user_id)

Gets a ban object for the given user from a guild.

get_guild_command(guild_id, command_id)
(since 0.6.0)

get_guild_command_permissions(guild_id)
(since 0.6.0)

get_guild_commands(guild_id)
(since 0.6.0)

get_guild_integrations(guild_id)

Gets a list of guild integerations.

Guild to get integrations for is specified by guild_id.

get_guild_member(guild_id, user_id)

Gets a guild member.

Examples

iex> Remedy.API.get_guild_member(4019283754613, 184937267485)

get_guild_prune_count(guild_id)

Gets the number of members that would be removed in a prune given days.

This endpoint requires the KICK_MEMBERS permission.

If successful, returns {:ok, %{pruned: pruned}}. Otherwise, returns a {:error, reason}.

Examples

iex> Remedy.API.get_guild_prune_count(81384788765712384, 1)
{:ok, %{pruned: 0}}

get_guild_sticker(guild_id, sticker_id)
(since 0.6.0)

Specs

get_guild_sticker(snowflake(), snowflake()) ::
  {:error, reason()} | {:ok, Remedy.Schema.Sticker.t()}

Gets a guild sticker by ID.

Examples

iex> Remedy.API.get_guild_sticker(guild_id, sticker_id)
{:ok, %Sticker{}}

iex Remedy.API.get_guild_sticker(bad_guild_id, sticker_id)
{:error, {404, 10060, "Unknown guild"}}

iex> Remedy.API.get_guild_sticker(guild_id, bad_sticker_id)
{:error, {404, 10060, "Unknown Sticker"}}

get_guild_template(template_code)
(since 0.6.0)

Specs

get_guild_template(any()) :: {:error, reason()} | {:ok, any()}

Get a guild template from the code or the full URL. eg: https://discord.new/2KAaMpa22ea6

Examples

iex>

get_guild_vanity_url(guild_id)
(since 0.6.0)

get_guild_welcome_screen(guild_id)
(since 0.6.0)

get_guild_widget(guild_id)
(since 0.6.0)

get_guild_widget_image(guild_id)
(since 0.6.0)

get_guild_widget_settings(guild_id)
(since 0.6.0)

get_invite(invite_code)

Gets an invite by its invite_code.

If successful, returns {:ok, invite}. Otherwise, returns a {:error, reason}.

Options

:with_counts (boolean) - whether to include member count fields

Examples

iex> Remedy.API.get_invite("zsjUsC")
{:ok, %Remedy.Schema.Invite{}}

iex> Remedy.API.get_invite("zsjUsC", with_counts: true)
{:ok, %Remedy.Schema.Invite{}}

get_message(channel_id, message_id)
(since 0.6.0)

Specs

get_message(snowflake(), snowflake()) ::
  {:error, any()} | {:ok, Remedy.Schema.Message.t()}

Retrieves a message from a channel.

Permissions

'VIEW_CHANNEL'

'READ_MESSAGE_HISTORY'

Examples

iex> Remedy.API.get_message(872417560094732331, 884355195277025321) {:ok, %Remedy.Schema.Message{}}

get_original_interaction_response(interaction_token)
(since 0.6.0)

get_reactions(channel_id, message_id, emoji)
(since 0.6.0)

Specs

get_reactions(snowflake(), snowflake(), any()) ::
  {:ok, [Remedy.Schema.User.t()]}

Gets all users who reacted with an emoji.

This endpoint requires the VIEW_CHANNEL and READ_MESSAGE_HISTORY permissions.

If successful, returns {:ok, users}. Otherwise, returns {:error, reason}.

See create_reaction/3 for similar examples.

get_stage(channel_id)
(since 0.6.0)

get_sticker(sticker_id)
(since 0.6.0)

Specs

get_sticker(snowflake()) ::
  {:error, any()}
  | {:ok,
     %Remedy.Schema.Sticker{
       __meta__: term(),
       asset: term(),
       available: term(),
       description: term(),
       format_type: term(),
       guild: term(),
       guild_id: term(),
       id: term(),
       name: term(),
       sort_value: term(),
       sticker_pack: term(),
       sticker_pack_id: term(),
       tags: term(),
       type: term(),
       user: term(),
       user_id: term()
     }}

Returns a sticker object.

Examples

iex> Remedy.API.get_sticker(818599312882794506)
{:ok, %Remedy.Schema.Sticker{}}

iex> Remedy.API.get_sticker(123)
{:error, {404, 10060, "Unknown sticker"}}

get_user(user_id)
(since 0.6.0)

Specs

get_user(snowflake()) :: {:error, reason()} | {:ok, Remedy.Schema.User.t()}

Get a user by id.

Examples

iex> Remedy.API.get_user(883307747305725972)
{:ok, %Remedy.Schema.User{
  id: 883307747305725972,
  avatar: "973a059282550a9ffaca42e795d8330b",
  username: "Remedy",
  ...
}}

iex Remedy.API.get_user(88330774730572597123)
{:error, {400, 50035, "Invalid Form Body"}}

get_webhook(webhook_id)

Gets a webhook by id.

Parameters

webhook_id - Id of the webhook to get.

get_webhook_message(webhook_id, webhook_token, message_id)
(since 0.6.0)

get_webhooks_with_token(webhook_id, webhook_token)

Gets a webhook by id and token.

This method is exactly like get_webhook/1 but does not require authentication.

Parameters

webhook_id - Id of the webhook to get.

webhook_token - Token of the webhook to get.

join_thread(channel_id)
(since 0.6.0)

Specs

join_thread(snowflake()) :: :ok | {:error, any()}

Adds the bot to the thread.

Events

:THREAD_MEMBERS_UPDATE

:THREAD_CREATE

Examples

iex> Remedy.API.join_thread(thread_the_bot_is_not_yet_in)
:ok

 iex> Remedy.API.join_thread(thread_the_bot_is_already_in)
:ok

iex> Remedy.API.join_thread(a_category_channel)
{:error, {400, 50024, "Cannot execute action on this channel type"}}

leave_guild(guild_id)

Specs

leave_guild(snowflake()) :: {:error, reason()} | :ok

Leaves a guild.

Examples

iex> Remedy.API.leave_guild(a_guild_i_dont_like)
:ok

iex> Remedy.API.leave_guild(a_guild_im_not_in)
:error, {400, 0, "400: Bad Request"}}

leave_thread(channel_id)
(since 0.6.0)

Specs

leave_thread(any()) :: :ok

Leaves a thread.

Also requires the thread is not archived.

Events

:THREAD_MEMBERS_UPDATE

Examples

iex> Remedy.API.leave_thread(channel_id)
:ok

list_active_channel_threads(channel_id)
(since 0.6.8)

Specs

list_active_channel_threads(Remedy.Schema.Channel.t()) ::
  {:error, reason()} | {:ok, [Remedy.Schema.Thread.t()]}

List the active threads in a channel.

Examples

iex> Remedy.API.list_active_channel_threads(a_valid_channel)
{:ok, [%Thread{}]}

list_active_guild_threads(guild_id)
(since 0.6.8)

Specs

list_active_guild_threads(Remedy.Schema.Guild.t()) ::
  {:error, reason()} | {:ok, [Remedy.Schema.Thread.t()]}

List the active threads in a guild.

Examples

iex> Remedy.API.list_active_guild_threads(a_valid_guild)
{:ok, [%Thread{}]}

list_active_threads(arg1)
(since 0.6.0)

Specs

list_active_threads(Remedy.Schema.Channel.t() | Remedy.Schema.Guild.t()) ::
  {:error, reason()} | {:ok, [Remedy.Schema.Thread.t()]}

List active threads.

This can be requested on both the /channels/ and the /guilds/ routes. To specify which route is used, you should pass a full %Guild{} or %Channel{} object.

list_channel_webhooks(channel_id)

Gets a list of webook for a channel.

Parameters

channel_id - Channel to get webhooks for.

list_emojis(guild_id)
(since 0.6.0)

Specs

list_emojis(snowflake()) :: {:error, reason()} | {:ok, [Emoji.t()]}

Gets a list of emojis for a given guild.

Permissions

:MANAGE_EMOJIS

Examples

iex> Remedy.API.list_emojis(guild_id)
{:ok, [%Emoji{}]}

iex> Remedy.API.list_emojis(bad_guild_id)
{:error, reason}

list_guild_bans(guild_id)
(since 0.6)

Gets a list of users banned from a guild.

Guild to get bans for is specified by guild_id.

list_guild_invites(guild_id)
(since 0.6)

Gets a list of invites for a guild.

This endpoint requires the MANAGE_GUILD permission.

If successful, returns {:ok, invites}. Otherwise, returns a {:error, reason}.

Examples

iex> Remedy.API.get_guild_invites(81384788765712384)
{:ok, [%Remedy.Schema.Invite{}]}

list_guild_members(guild_id)

Gets a list of a guild's members.

Options

:limit (integer) - max number of members to return (1-1000) (default: 1)

:after (t:Remedy.Schema.User.id/0) - the highest user id in the previous page (default: 0)

Examples

iex>  Remedy.API.list_guild_members(41771983423143937, limit: 1)

list_guild_roles(guild_id)
(since 0.6.0)

Specs

list_guild_roles(snowflake()) ::
  {:ok, [Remedy.Schema.Role.t()]} | {:error, reason()}

Gets a guild's roles.

Examples

iex>  Remedy.API.get_guild_roles(147362948571673)
{:ok, [%Remedy.Schema.Role{}]}

list_guild_stickers(guild_id)
(since 0.6.0)

Specs

list_guild_stickers(snowflake()) ::
  {:error, reason()} | {:ok, [Remedy.Schema.Sticker.t()]}

List all custom stickers for a guild id.

Examples

iex> Remedy.API.list_guild_stickers(guild_id)
{:ok, [%Sticker{}]}

iex> Remedy.API.get_sticker(123)
{:error, {404, 10060, "Unknown guild"}}

list_guild_voice_regions(guild_id)

Gets a list of voice regions for the guild.

Guild to get voice regions for is specified by guild_id.

list_guild_webhooks(guild_id)

Gets a list of webooks for a guild.

Parameters

guild_id - Guild to get webhooks for.

list_guilds(params)

Gets a list of guilds the bot is currently in.

Options

:before, Snowflake

:after, Snowflake

:limit, :integer, min: 1, max: 100

Examples

iex> Remedy.API.list_guilds(limit: 1)
{:ok, [%Remedy.Schema.Guild{}]}

list_guilds_opts()

list_joined_private_archived_threads(channel_id)
(since 0.6.0)

Specs

list_joined_private_archived_threads(snowflake()) ::
  {:ok, [Remedy.Schema.Thread.t()]} | {:error, reason()}

list_messages(channel_id, params \\ [])
(since 0.6)

Specs

list_messages(snowflake(), opts()) ::
  {:error, any()} | {:ok, [Remedy.Schema.Channel.t()]}

Retrieves a channel's messages.

Permissions

VIEW_CHANNEL

READ_MESSAGE_HISTORY

Options

:before*

:after*

:around*

:limit

Only one may be provided.

Helpers

Remedy.API.list_messages_before/3

Remedy.API.list_messages_after/3

Remedy.API.list_messages_around/3

Examples

iex> Remedy.API.list_messages(872417560094732331, [{:before, 882781809908256789}, {:limit, 1}])
{:ok, [%Message{id: 882681855315423292}]}

list_messages_after(channel_id, message_id, limit \\ 50)
(since 0.6.0)

Specs

list_messages_after(snowflake(), snowflake(), limit()) ::
  {:error, any()} | {:ok, list()}

List messages after a given message.

Permissions

VIEW_CHANNEL

READ_MESSAGE_HISTORY

list_messages_around(channel_id, message_id, limit \\ 50)

Specs

list_messages_around(snowflake(), snowflake(), limit()) ::
  {:error, any()} | {:ok, list()}

List messages around a given message.

Permissions

VIEW_CHANNEL

READ_MESSAGE_HISTORY

list_messages_before(channel_id, message_id, limit \\ 50)
(since 0.6.0)

Specs

list_messages_before(snowflake(), snowflake(), limit()) ::
  {:error, any()} | {:ok, list()}

List messages before a given message.

Permissions

VIEW_CHANNEL

READ_MESSAGE_HISTORY

list_nitro_sticker_packs()
(since 0.6.0)

Specs

list_nitro_sticker_packs() ::
  {:error, reason()} | {:ok, [Remedy.Schema.Sticker.t()]}

List the Nitro Sticker Packs

list_pinned_messages(channel_id)
(since 0.6.0)

Retrieves all pinned messages from a channel.

Permissions

'VIEW_CHANNEL'

'READ_MESSAGE_HISTORY'

Examples

iex> Remedy.API.get_pinned_messages(43189401384091)

list_private_archived_threads(channel_id)
(since 0.6.0)

Specs

list_private_archived_threads(snowflake()) ::
  {:ok, [Remedy.Schema.Thread.t()]} | {:error, reason()}

List private archived threads in the given channel.

Examples

iex> Remedy.API.list_public_archived_threads(channel_id) {:ok, [%Thread{}]}

list_public_archived_threads(channel_id)
(since 0.6.0)

Specs

list_public_archived_threads(snowflake()) :: {:error, any()} | {:ok, any()}

List public archived threads in the given channel.

Examples

iex> Remedy.API.list_public_archived_threads(channel_id)
{:ok, [%Thread{}]}

list_thread_members(channel_id)
(since 0.6.0)

Specs

list_thread_members(snowflake()) ::
  {:error, reason()} | {:ok, [Remedy.Schema.User.t()]}

list_voice_regions()

Gets a list of voice regions.

Examples

iex> Remedy.API.list_voice_regions()
{:ok,
[
  %{
    custom: false,
    deprecated: false,
    id: "us-west",
    name: "US West",
    optimal: false
  }
]

list_webhooks(guild_id)
(since 0.6.0)

modify_bot(opts)
(since 0.6.0)

Specs

modify_bot(opts()) :: {:error, reason()} | {:ok, Remedy.Schema.User.t()}

Changes the attributes of the bot.

Options

:username, :string

:avatar, :string see: avatar data

Examples

iex> Remedy.API.modify_current_user(avatar: "data:image/jpeg;base64,YXl5IGJieSB1IGx1a2luIDQgc3VtIGZ1az8=")
{:ok, %Remedy.Schema.User{}}

modify_bot_opts()

modify_channel(channel)
(since 0.6.8)

Specs

modify_channel(Remedy.Schema.Channel.t()) ::
  {:error, any()} | {:ok, Remedy.Schema.Channel.t()}

Modify a channel's settings with a channel object.

Examples

iex> Remedy.API.modify_channel(%Channel{})
{:ok, %Channel{}}

modify_channel(channel_id, opts)
(since 0.6.0)

Specs

modify_channel(snowflake(), opts()) ::
  {:error, any()} | {:ok, Remedy.Schema.Channel.t()}

Modifies a channel's settings.

Permissions

MANAGE_CHANNEL

Events

:CHANNEL_UPDATE.

Options

:name - :string, min: 2, max: 100

:position - :integer - Not contiguous, will be ordered by id if duplicates exist.

:topic :string, mix: 0, max: 1024

:nsfw :boolean, default: false

:bitrate - :integer, min: 8000, max: 128000

:user_limit - :integer, min: 1, max: 99, unlimited: 0

:permission_overwrites - {:array, %Overwrite{}}

:parent_id - category to place the channel under.

:reason - reason for the audit log.

Examples

iex> Remedy.API.modify_channel(41771983423143933, name: "elixir-remedy", topic: "remedy discussion") {:ok, %Remedy.Schema.Channel{id: 41771983423143933, name: "elixir-remedy", topic: "remedy discussion"}}

modify_channel_permissions(channel_id, overwrite_id, opts, reason)

Specs

modify_channel_permissions(
  any(),
  any(),
  {:error, any()} | {:ok, any()} | map(),
  any()
) :: {:error, any()} | {:ok, any()}

Edit the permission overwrites for a user or role.

Role or user to overwrite is specified by overwrite_id.

permission_info is a map with the following keys:

type - Required; member if editing a user, role if editing a role.

allow - Bitwise value of allowed permissions.

deny - Bitwise value of denied permissions.

type - member if editing a user, role if editing a role.

An optional reason can be provided for the audit log.

allow and deny are defaulted to 0, meaning that even if you don't specify them, they will override their respective former values in an existing overwrite.

modify_channel_positions(guild_id, opts)

Reorders a guild's channels.

Permissions

MANAGE_CHANNELS

Events

:CHANNEL_UPDATE

Options

positions

Examples

iex> Remedy.API.modify_channel_positions(279093381723062272, [%{id: 351500354581692420, position: 2}])
{:ok}

iex> Remedy.API.modify_channel_positions(279093381723062272, [%{id: 351500354581692420, position: 2}])
{:ok}

modify_channel_positions_opts()

modify_command_permissions(guild_id)
(since 0.6.0)

modify_command_permissions(guild_id, command_id)
(since 0.6.0)

modify_current_user_voice_state(guild_id)
(since 0.6.0)

modify_emoji(guild_id, emoji_id, opts)
(since 0.6.0)

Specs

modify_emoji(snowflake(), snowflake(), opts()) :: {:ok, Emoji.t()}

Modify the given emoji.

Permissions

:MANAGE_EMOJIS

Events

:EMOJIS_UPDATE

Options

:name - :string

:roles - [:role]

:reason - :string

Examples

iex> Remedy.API.modify_emoji(43189401384091, 4314301984301, name: "elixir", roles: [])
{:ok, %Remedy.Schema.Emoji{}}

modify_followup(interaction_token, message_id)
(since 0.6.0)

modify_global_command(command_id)

Update an existing global application command.

The updated command will be available on all guilds in around an hour.

Parameters

application_id: Application ID for which to edit the command. If not given, this will be fetched from Me.

command_id: The current snowflake of the command.

command: Command configuration, see the linked API documentation for reference.

Return value

The updated command. See the official reference: https://discord.com/developers/docs/interactions/slash-commands#edit-global-application-command

modify_guild(guild_id, opts)

Modifies a guild's settings.

Permissions

MANAGE_GUILD

Events

:GUILD_UPDATE

Options

:name - :string, min: 2, max: 100

:icon - :image_data

:verification_level - :integer

0 - :none - unrestricted

1 - :low - must have verified email on account

2 - :medium - must be registered on Discord for longer than 5 minutes

3 - :high - must be a member of the server for longer than 10 minutes

4 - :very_high - must have a verified phone number

:default_message_notifications - :integer

0 - :all_messages - members will receive notifications for all messages by default

1 - :only_mentions - members will receive notifications only for mentions by default

:explicit_content_filter - :integer

0 - :disabled

1 - :members_without_roles

2 - :all_members

:roles - {:array, :role}

:channels - {:array, :channel}

:afk_channel_id - :snowflake

:afk_timeout - :integer, :seconds

:icon - :string

:owner_id - Snowflake - to transfer guild ownership to (must be owner)

:splash - :string

:system_channel_id - Snowflake

:rules_channel_id - Snowflake

:public_updates_channel_id - Snowflake

Audit Log

:reason

Examples

iex> Remedy.API.modify_guild(451824027976073216, name: "Nose Drum")
{:ok, %Remedy.Schema.Guild{id: 451824027976073216, name: "Nose Drum", ...}}

modify_guild_command(guild_id, command_id)
(since 0.6.0)

modify_guild_member(guild_id, user_id, opts)

Modifies a guild member's attributes.

Permissions

MANAGE_NICKNAMES

MANAGE_ROLES

MUTE_MEMBERS

DEAFEN_MEMBERS

MOVE_MEMBERS

Events

:GUILD_MEMBER_UPDATE

Options

:nick (string) - value to set users nickname to

:roles - array of role ids the member is assigned

:mute (boolean) - if the user is muted

:deaf (boolean) - if the user is deafened

:channel_id - id of channel to move user to (if they are connected to voice)

Examples

iex> Remedy.API.modify_guild_member(41771983423143937, 637162356451, nick: "Remedy")
{:ok}

modify_guild_member_opts()

modify_guild_role(guild_id, role_id)

Modifies a guild role.

This endpoint requires the MANAGE_ROLES permission. It fires a Remedy.Consumer.guild_role_update/0 event.

An optional reason can be specified for the audit log.

If successful, returns {:ok, role}. Otherwise, returns a {:error, reason}.

Options

:name (string) - name of the role

:permissions (integer) - bitwise of the enabled/disabled permissions

:color (integer) - RGB color value (default: 0)

:hoist (boolean) - whether the role should be displayed separately in the sidebar

:mentionable (boolean) - whether the role should be mentionable

Examples

iex> Remedy.API.modify_guild_role(41771983423143937, 392817238471936, hoist: false, name: "foo-bar")

modify_guild_role_positions(guild_id)

Reorders a guild's roles.

This endpoint requires the MANAGE_ROLES permission. It fires multiple Remedy.Consumer.guild_role_update/0 events.

If successful, returns {:ok, roles}. Otherwise, returns a {:error, reason}.

positions is a list of maps that each map a role id with a position.

Examples

iex> Remedy.API.modify_guild_role_positions(41771983423143937, [%{id: 41771983423143936, position: 2}])

modify_guild_sticker(guild_id, sticker_id)
(since 0.6.0)

modify_guild_template(guild_id, template_code)
(since 0.6.0)

modify_guild_welcome_screen(guild_id)
(since 0.6.0)

modify_message(channel_id, message_id, opts)
(since 0.6.0)

Edits a previously sent message in a channel.

Permissions

VIEW_CHANNEL

Events

:MESSAGE_UPDATE.

Options

:content (string) - the message contents (up to 2000 characters)

:embed (Remedy.Schema.Embed.t/0) - embedded rich content

Examples

iex> Remedy.API.modify_message(889614079830925352, 1894013840914098, content: "hello world!")
:ok

iex> Remedy.API.modify_message(889614079830925352, 1894013840914098, "hello world!")
:ok

modify_nickname(guild_id, nickname)

Modifies the nickname of the current user in a guild.

Options

:nick (string) - value to set users nickname to

Examples

iex> Remedy.API.modify_nickname(41771983423143937, nick: "Remedy")
{:ok, %{nick: "Remedy"}}

iex>

modify_original_interaction_response(interaction_token)
(since 0.6.0)

modify_stage(channel_id)
(since 0.6.0)

modify_user_voice_state(guild_id, user_id)
(since 0.6.0)

modify_webhook(webhook_id)

Modifies a webhook.

Parameters

webhook_id - Id of the webhook to modify.

args - Map with the following optional keys:

name - Name of the webhook.

avatar - Base64 128x128 jpeg image for the default avatar.

reason - An optional reason for the guild audit log.

modify_webhook_message(webhook_id, webhook_token, message_id)
(since 0.6.0)

modify_webhooks_with_token(webhook_id, webhook_token)

Modifies a webhook with a token.

This method is exactly like modify_webhook/1 but does not require authentication.

Parameters

webhook_id - Id of the webhook to modify.

webhook_token - Token of the webhook to get.

args - Map with the following optional keys:

name - Name of the webhook.

avatar - Base64 128x128 jpeg image for the default avatar.

reason - An optional reason for the guild audit log.

overwrite_global_commands(opts)

Overwrite the existing global application commands.

This action will:

Create any command that was provided and did not already exist

Update any command that was provided and already existed if its configuration changed

Delete any command that was not provided but existed on Discord's end

Updates will be available in all guilds after 1 hour. Commands that do not already exist will count toward daily application command create limits.

Parameters

commands: List of command configurations, see the linked API documentation for reference.

Return value

Updated list of global application commands. See the official reference: https://discord.com/developers/docs/interactions/slash-commands#bulk-overwrite-global-application-commands

overwrite_guild_commands(guild_id)

Overwrite the existing guild application commands on the specified guild.

This action will:

Create any command that was provided and did not already exist

Update any command that was provided and already existed if its configuration changed

Delete any command that was not provided but already exists

This is functionally similar to Ecto's cast_assoc

opts

guild_id: Guild on which to overwrite the commands.

commands: List of command configurations, see the linked API documentation for reference.

Return value

Updated list of guild application commands. See the official reference: https://discord.com/developers/docs/interactions/slash-commands#bulk-overwrite-guild-application-commands

pin_message(channel_id, message_id)

Specs

pin_message(any(), any()) :: :ok | {:error, any()}

Pins a message in a channel.

Permissions

'VIEW_CHANNEL'

'READ_MESSAGE_HISTORY'

'MANAGE_MESSAGES'

Events

:MESSAGE_UPDATE

:CHANNEL_PINS_UPDATE

Examples

iex> Remedy.API.pin_message(43189401384091, 18743893102394) :ok

publish_message(message)
(since 0.6.0)

Specs

publish_message(Remedy.Schema.Message.t()) :: {:error, any()} | {:ok, any()}

Publish a message in a news channel.

This will propagate a message out to all followers of the channel.

This is known as "Crosspost Message" in the Discord API.

publish_message(channel_id, message_id)
(since 0.6.8)

Specs

publish_message(snowflake(), snowflake()) :: {:error, any()} | {:ok, map()}

remove_guild_ban(guild_id, user_id, opts)
(since 0.6.8)

Specs

remove_guild_ban(snowflake(), snowflake(), opts()) :: :ok | {:error, reason()}

Removes a ban for a user.

Options

:reason - Reason for the audit log.

Examples

iex> Remedy.API.remove_guild_ban(guild_id, user_id)
:ok

remove_guild_member(guild_id, user_id)

Removes a member from a guild.

This event requires the KICK_MEMBERS permission. It fires a Remedy.Consumer.guild_member_remove/0 event.

An optional reason can be provided for the audit log with reason.

If successful, returns {:ok}. Otherwise, returns a {:error, reason}.

Examples

iex> Remedy.API.remove_guild_member(1453827904102291, 18739485766253)
{:ok}

remove_guild_member_role(guild_id, user_id, role_id)

Removes a role from a member.

Role to remove is specified by role_id. User to remove role from is specified by guild_id and user_id. An optional reason can be given for the audit log.

remove_thread_member(channel_id, user_id)
(since 0.6.0)

Specs

remove_thread_member(snowflake(), snowflake()) :: :ok | {:error, any()}

Remove a user from a thread.

Also requires the thread is not archived.

Events

:THREAD_MEMBERS_UPDATE

Examples

iex> Remedy.API.remove_thread_member(channel_id, user_id) :ok

search_guild_members(guild_id)
(since 0.6.0)

start_thread_with_message(channel_id, message_id, opts)
(since 0.6.0)

Starts a thread with a message.

Events

start_thread_with_message_opts()

start_thread_without_message(channel_id, opts)
(since 0.6.0)

Specs

start_thread_without_message(snowflake(), opts()) ::
  {:error, reason()} | {:ok, Remedy.Schema.Thread.t()}

start_thread_without_message_opts()

start_typing(channel_id)
(since 0.6.8)

Specs

start_typing(snowflake()) :: :ok | {:error, reason()}

Triggers the typing indicator.

Events

:TYPING_START

Examples

iex> Remedy.API.start_typing(891925736120791080)
:ok

sync_guild_from_template(guild_id, template_code)
(since 0.6.0)

unpin_message(channel_id, message_id)

Specs

unpin_message(snowflake(), snowflake()) :: :ok | {:error, reason()}

Unpins a message in a channel.

Permissions

'VIEW_CHANNEL'

'READ_MESSAGE_HISTORY'

'MANAGE_MESSAGES'

Events

:MESSAGE_UPDATE

:CHANNEL_PINS_UPDATE

Examples

iex> Remedy.API.unpin_message(43189401384091, 18743893102394) :ok