Remedy.Api (Remedy v0.5.4)

Hello

Types

emoji()

Specs

emoji() :: Remedy.Schema.Emoji.t() | Remedy.Schema.Emoji.api_name()

error()

Specs

error() :: {:error, Remedy.ApiError.t() | HTTPoison.Error.t()}

limit()

Specs

limit() :: integer() | :infinity

locator()

Specs

locator() ::
  {:before, integer()} | {:after, integer()} | {:around, integer()} | {}

matrix()

Specs

matrix() :: m1() | m2() | m3()

options()

Specs

options() :: keyword() | map()

status()

Specs

status() :: :dnd | :idle | :online | :invisible

Functions

add_guild_member(guild_id, user_id, options)

Specs

add_guild_member(Remedy.Schema.Guild.id(), Remedy.Schema.User.id(), options()) ::
  error() | {:ok, Remedy.Schema.Member.t()} | {:ok}

Puts a user in a guild.

This endpoint fires the t:Remedy.Consumer.guild_member_add/0 event. It requires the CREATE_INSTANT_INVITE permission. Additionally, it situationally requires the MANAGE_NICKNAMES, MANAGE_ROLES, MUTE_MEMBERS, and DEAFEN_MEMBERS permissions.

If successful, returns {:ok, member} or {:ok} if the user was already a member of the guild. Otherwise, returns a Remedy.Api.error/0.

Options

:access_token (string) - the user's oauth2 access token

:nick (string) - value to set users nickname to

:roles (list of t:Remedy.Struct.Guild.Role.id/0) - array of role ids the member is assigned

:mute (boolean) - if the user is muted

:deaf (boolean) - if the user is deafened

:access_token is always required.

Examples

iex>

Remedy.Api.add_guild_member( 41771983423143937, 18374719829378473, access_token: "6qrZcUqja7812RVdnEKjpzOL4CvHBFG", nick: "remedy", roles: [431849301, 913809431] )

add_guild_member_role(guild_id, user_id, role_id, reason \\ nil)

Specs

add_guild_member_role(
  integer(),
  integer(),
  integer(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok}

Adds a role to a member.

Role to add is specified by role_id. User to add role to is specified by guild_id and user_id. An optional reason can be given for the audit log.

add_pinned_channel_message(channel_id, message_id)

Specs

add_pinned_channel_message(
  Remedy.Schema.Channel.id(),
  Remedy.Schema.Message.id()
) :: error() | {:ok}

Pins a message in a channel.

This endpoint requires the 'VIEW_CHANNEL', 'READ_MESSAGE_HISTORY', and 'MANAGE_MESSAGES' permissions. It fires the t:Remedy.Consumer.message_update/0 and t:Remedy.Consumer.channel_pins_update/0 events.

If successful, returns {:ok}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.add_pinned_channel_message(43189401384091, 18743893102394)

begin_guild_prune(guild_id, days, reason \\ nil)

Specs

begin_guild_prune(
  Remedy.Schema.Guild.id(),
  1..30,
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok, %{pruned: integer()}}

Begins a guild prune to prune members within days.

An optional reason can be provided for the guild audit log.

This endpoint requires the KICK_MEMBERS permission. It fires multiple t:Remedy.Consumer.guild_member_remove/0 events.

If successful, returns {:ok, %{pruned: pruned}}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.begin_guild_prune(81384788765712384, 1) {:ok, %{pruned: 0}}

bulk_delete_messages(channel_id, messages, filter \\ true)

Specs

bulk_delete_messages(integer(), [Remedy.Struct.Message.id()], boolean()) ::
  error() | {:ok}

Deletes multiple messages from a channel.

messages is a list of Remedy.Struct.Message.id that you wish to delete. When given more than 100 messages, this function will chunk the given message list into blocks of 100 and send them off to the API. It will stop deleting on the first error that occurs. Keep in mind that deleting thousands of messages will take a pretty long time and it may be proper to just delete the channel you want to bulk delete in and recreate it.

This method can only delete messages sent within the last two weeks. Filter is an optional parameter that specifies whether messages sent over two weeks ago should be filtered out; defaults to true.

bulk_overwrite_global_application_commands(application_id \\ Cache.bot().id, commands)
(since 0.5.0)

Specs

bulk_overwrite_global_application_commands(Remedy.Schema.User.id(), [map()]) ::
  {:ok, [map()]} | error()

Overwrite the existing global application commands.

This action will:

Create any command that was provided and did not already exist

Update any command that was provided and already existed if its configuration changed

Delete any command that was not provided but existed on Discord's end

Updates will be available in all guilds after 1 hour. Commands that do not already exist will count toward daily application command create limits.

Parameters

application_id: Application ID for which to overwrite the commands. If not given, this will be fetched from Me.

commands: List of command configurations, see the linked API documentation for reference.

Return value

Updated list of global application commands. See the official reference: https://discord.com/developers/docs/interactions/slash-commands#bulk-overwrite-global-application-commands

bulk_overwrite_guild_application_commands(application_id \\ Cache.bot().id, guild_id, commands)
(since 0.5.0)

Specs

bulk_overwrite_guild_application_commands(
  Remedy.Schema.User.id(),
  Remedy.Schema.Guild.id(),
  [map()]
) :: {:ok, [map()]} | error()

Overwrite the existing guild application commands on the specified guild.

This action will:

Create any command that was provided and did not already exist

Update any command that was provided and already existed if its configuration changed

Delete any command that was not provided but existed on Discord's end

Parameters

application_id: Application ID for which to overwrite the commands. If not given, this will be fetched from Me.

guild_id: Guild on which to overwrite the commands.

commands: List of command configurations, see the linked API documentation for reference.

Return value

Updated list of guild application commands. See the official reference: https://discord.com/developers/docs/interactions/slash-commands#bulk-overwrite-guild-application-commands

create_channel_invite(channel_id, options \\ [], reason \\ nil)

Specs

create_channel_invite(
  Remedy.Schema.Channel.id(),
  options(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok, Invite.detailed_invite()}

Creates an invite for a guild channel.

An optional reason can be provided for the audit log.

This endpoint requires the CREATE_INSTANT_INVITE permission.

If successful, returns {:ok, invite}. Otherwise, returns a Remedy.Api.error/0.

Options

:max_age (integer) - duration of invite in seconds before expiry, or 0 for never. (default: 86400)

:max_uses (integer) - max number of uses or 0 for unlimited. (default: 0)

:temporary (boolean) - Whether the invite should grant temporary membership. (default: false)

:unique (boolean) - used when creating unique one time use invites. (default: false)

Examples

iex>

Remedy.Api.create_channel_invite(41771983423143933) {:ok, Remedy.Struct.Invite{}}

Remedy.Api.create_channel_invite(41771983423143933, max_uses: 20) {:ok, %Remedy.Struct.Invite{}}

create_dm(user_id)

Specs

create_dm(Remedy.Schema.User.id()) ::
  error() | {:ok, Remedy.Schema.Channel.dm_channel()}

Create a new DM channel with a user.

If successful, returns {:ok, dm_channel}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.create_dm(150061853001777154) {:ok, %Remedy.Struct.Channel{type: 1}}

create_followup_message(application_id \\ Cache.bot().id, token, webhook_payload)

Specs

create_followup_message(
  Remedy.Schema.User.id(),
  Remedy.Schema.Interaction.token(),
  map()
) :: {:ok} | error()

Create a followup message for an interaction.

Delegates to execute_webhook/3, see the function for more details.

create_global_application_command(application_id \\ Cache.bot().id, command)

Specs

create_global_application_command(Remedy.Schema.User.id(), map()) ::
  {:ok, map()} | error()

Create a new global application command.

The new command will be available on all guilds in around an hour. If you want to test commands, use create_guild_application_command/2 instead, as commands will become available instantly there. If an existing command with the same name exists, it will be overwritten.

Parameters

application_id: Application ID for which to create the command. If not given, this will be fetched from Me.

command: Command configuration, see the linked API documentation for reference.

Return value

The created command. See the official reference: https://discord.com/developers/docs/interactions/slash-commands#create-global-application-command

Example

iex>

Remedy.Api.create_application_command( %{name: "edit", description: "ed, man! man, ed", options: []} )

create_group_dm(access_tokens, nicks)

Specs

create_group_dm([String.t()], %{optional(Remedy.Schema.User.id()) => String.t()}) ::
  error() | {:ok, Remedy.Schema.Channel.group_dm_channel()}

Creates a new group DM channel.

If successful, returns {:ok, group_dm_channel}. Otherwise, returns a Remedy.Api.error/0.

access_tokens are user oauth2 tokens. nicks is a map that maps a user id to a nickname.

Examples

iex>

Remedy.Api.create_group_dm(["6qrZcUqja7812RVdnEKjpzOL4CvHBFG"], %{41771983423143937 => "My Nickname"}) {:ok, %Remedy.Struct.Channel{type: 3}}

create_guild_application_command(application_id \\ Cache.bot().id, guild_id, command)

Specs

create_guild_application_command(
  Remedy.Schema.User.id(),
  Remedy.Schema.Guild.id(),
  map()
) :: {:ok, map()} | error()

create_guild_ban(guild_id, user_id, days_to_delete, reason \\ nil)

Specs

create_guild_ban(
  integer(),
  integer(),
  integer(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok}

Bans a user from a guild.

User to delete is specified by guild_id and user_id. An optional reason can be specified for the audit log.

create_guild_channel(guild_id, options)

Specs

create_guild_channel(Remedy.Schema.Guild.id(), options()) ::
  error() | {:ok, Remedy.Schema.Channel.guild_channel()}

Creates a channel for a guild.

This endpoint requires the MANAGE_CHANNELS permission. It fires a t:Remedy.Consumer.channel_create/0 event.

If successful, returns {:ok, channel}. Otherwise, returns a Remedy.Api.error/0.

Options

:name (string) - channel name (2-100 characters)

:type (integer) - the type of channel (See Remedy.Struct.Channel)

:topic (string) - channel topic (0-1024 characters)

:bitrate (integer) - the bitrate (in bits) of the voice channel (voice only)

:user_limit (integer) - the user limit of the voice channel (voice only)

:permission_overwrites (list of t:Remedy.Struct.Overwrite.t/0 or equivalent map) - the channel's permission overwrites

:parent_id (t:Remedy.Struct.Channel.id/0) - id of the parent category for a channel

:nsfw (boolean) - if the channel is nsfw

:name is always required.

Examples

iex>

Remedy.Api.create_guild_channel(81384788765712384, name: "elixir-remedy", topic: "craig's domain") {:ok, %Remedy.Struct.Channel{guild_id: 81384788765712384}}

create_guild_emoji(guild_id, options, reason \\ nil)

Specs

create_guild_emoji(
  Remedy.Schema.Guild.id(),
  options(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok, Remedy.Schema.Emoji.t()}

Creates a new emoji for the given guild.

This endpoint requires the MANAGE_EMOJIS permission. It fires a t:Remedy.Consumer.guild_emojis_update/0 event.

An optional reason can be provided for the audit log.

If successful, returns {:ok, emoji}. Otherwise, returns Remedy.Api.error/0.

Options

:name (string) - name of the emoji

:image (base64 data URI) - the 128x128 emoji image. Maximum size of 256kb

:roles (list of t:Sunbake.Snowflake.t/0) - roles for which this emoji will be whitelisted (default: [])

:name and :image are always required.

Examples

iex>

image = "data:image/png;base64,YXl5IGJieSB1IGx1a2luIDQgc3VtIGZ1az8="

Remedy.Api.create_guild_emoji(43189401384091, name: "remedy", image: image, roles: [])

create_guild_integrations(guild_id, options)

Specs

create_guild_integrations(integer(), %{type: String.t(), id: integer()}) ::
  error() | {:ok}

Creates a new guild integeration.

Guild to create integration with is specified by guild_id.

options is a map with the following requires keys:

type - Integration type.

id - Integeration id.

create_guild_role(guild_id, options, reason \\ nil)

Specs

create_guild_role(
  Remedy.Schema.Guild.id(),
  options(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok, Remedy.Schema.Role.t()}

Creates a guild role.

An optional reason for the audit log can be provided via reason.

This endpoint requires the MANAGE_ROLES permission. It fires a t:Remedy.Consumer.guild_role_create/0 event.

If successful, returns {:ok, role}. Otherwise, returns a Remedy.Api.error/0.

Options

:name (string) - name of the role (default: "new role")

:permissions (integer) - bitwise of the enabled/disabled permissions (default: @everyone perms)

:color (integer) - RGB color value (default: 0)

:hoist (boolean) - whether the role should be displayed separately in the sidebar (default: false)

:mentionable (boolean) - whether the role should be mentionable (default: false)

Examples

iex>

Remedy.Api.create_guild_role(41771983423143937, name: "remedy-club", hoist: true)

create_interaction_response(interaction, response)

Specs

create_interaction_response(Remedy.Schema.Interaction.t(), map()) ::
  {:ok} | error()

Same as create_interaction_response/3, but directly takes the t:Remedy.Struct.Interaction.t/0 received from the gateway.

create_interaction_response(id, token, response)

Specs

create_interaction_response(
  Remedy.Schema.Interaction.id(),
  Remedy.Schema.Interaction.token(),
  map()
) :: {:ok} | error()

Create a response to an interaction received from the gateway.

Parameters

id: The interaction ID to which the response should be created.

token: The interaction token.

response: An InteractionResponse object. See the linked documentation.

Example

iex>

response = %{ type: 4, data: %{

content: "I copy and pasted this code."

} } Remedy.Api.create_interaction_response(interaction, response)

As an alternative to passing the interaction ID and token, the original t:Remedy.Struct.Interaction.t/0 can also be passed directly. See create_interaction_response/1.

create_message(channel_id, options)

Specs

create_message(
  Remedy.Schema.Channel.id() | Remedy.Schema.Message.t(),
  options() | String.t()
) :: error() | {:ok, Remedy.Schema.Message.t()}

Posts a message to a guild text or DM channel.

Intents

:VIEW_CHANNEL

:SEND_MESSAGES

:SEND_MESSAGES_TTS (optional)

Events

t:Remedy.Gateway.Dispatch.message_create/0.

Options

:content (string) - the message contents (up to 2000 characters)

:nonce (t:Sunbake.Snowflake.t/0) - a nonce that can be used for optimistic message sending

:tts (boolean) - true if this is a TTS message

:file (Path.t/0 | map) - the path of the file being sent, or a map with the following keys if sending a binary from memory

:name (string) - the name of the file

:body (string) - binary you wish to send

:embed (t:Remedy.Struct.Embed.t/0) - embedded rich content

:allowed_mentions - See "Allowed mentions" below

:message_reference (map) - See "Message references" below

Note: At least one of the following is required: :content, :file, :embed.

Allowed mentions

:all (default) - Ping everything as usual

:none - Nobody will be pinged

:everyone - Allows to ping @here and @everone

:users - Allows to ping users

:roles - Allows to ping roles

{:users, list} - Allows to ping list of users. Can contain up to 100 ids of users.

{:roles, list} - Allows to ping list of roles. Can contain up to 100 ids of roles.

Examples

iex> Remedy.Api.create_message(872417560094732331, content: "hello world!")
{:ok, %Message{}}

create_reaction(channel_id, message_id, emoji)

Specs

create_reaction(Remedy.Schema.Channel.id(), Remedy.Schema.Message.id(), emoji()) ::
  error() | {:ok}

Creates a reaction for a message.

This endpoint requires the VIEW_CHANNEL and READ_MESSAGE_HISTORY permissions. Additionally, if nobody else has reacted to the message with the emoji, this endpoint requires the ADD_REACTIONS permission. It fires a t:Remedy.Consumer.message_reaction_add/0 event.

If successful, returns {:ok}. Otherwise, returns Remedy.Api.error/0.

Examples

iex>

Using a Remedy.Struct.Emoji.

emoji = %Remedy.Struct.Emoji{id: 43819043108, name: "foxbot"} Remedy.Api.create_reaction(123123123123, 321321321321, emoji)

Using a base 16 emoji string.

Remedy.Api.create_reaction(123123123123, 321321321321, "ðŸ˜")

For other emoji string examples, see t:Remedy.Struct.Emoji.api_name/0.

create_webhook(channel_id, args, reason \\ nil)

Specs

create_webhook(
  Remedy.Schema.Channel.id(),
  %{name: String.t(), avatar: String.t()},
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok, Remedy.Struct.Webhook.t()}

Creates a webhook.

Parameters

channel_id - Id of the channel to send the message to.

args - Map with the following required keys:

name - Name of the webhook.

avatar - Base64 128x128 jpeg image for the default avatar.

reason - An optional reason for the guild audit log.

delete_all_reactions(channel_id, message_id)

Specs

delete_all_reactions(Remedy.Schema.Channel.id(), Remedy.Schema.Message.id()) ::
  error() | {:ok}

Deletes all reactions from a message.

This endpoint requires the VIEW_CHANNEL, READ_MESSAGE_HISTORY, and MANAGE_MESSAGES permissions. It fires a t:Remedy.Consumer.message_reaction_remove_all/0 event.

If successful, returns {:ok}. Otherwise, return Remedy.Api.error/0.

delete_channel(channel_id, reason \\ nil)

Specs

delete_channel(Remedy.Schema.Channel.id(), Remedy.Schema.AuditLogEntry.reason()) ::
  error() | {:ok, Remedy.Schema.Channel.t()}

Deletes a channel.

An optional reason can be provided for the guild audit log.

If deleting a t:Remedy.Struct.Channel.guild_channel/0, this endpoint requires the MANAGE_CHANNELS permission. It fires a t:Remedy.Consumer.channel_delete/0. If a t:Remedy.Struct.Channel.channel_category/0 is deleted, then a t:Remedy.Consumer.channel_update/0 event will fire for each channel under the category.

If successful, returns {:ok, channel}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex> Remedy.Api.delete_channel(421533712753360896)
{:ok, %Remedy.Struct.Channel{id: 421533712753360896}}

delete_channel_permissions(channel_id, overwrite_id, reason \\ nil)

Specs

delete_channel_permissions(
  integer(),
  integer(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok}

Delete a channel permission for a user or role.

Role or user overwrite to delete is specified by channel_id and overwrite_id. An optional reason can be given for the audit log.

delete_global_application_command(application_id \\ Cache.bot().id, command_id)

Specs

delete_global_application_command(Remedy.Schema.User.id(), Snowflake.t()) ::
  {:ok} | error()

Delete an existing global application command.

Parameters

application_id: Application ID for which to create the command. If not given, this will be fetched from Me.

command_id: The current snowflake of the command.

delete_guild(guild_id)

Specs

delete_guild(Remedy.Schema.Guild.id()) :: error() | {:ok}

Deletes a guild.

This endpoint requires that the current user is the owner of the guild. It fires the t:Remedy.Consumer.guild_delete/0 event.

If successful, returns {:ok}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.delete_guild(81384788765712384)

delete_guild_application_command(application_id \\ Cache.bot().id, guild_id, command_id)

Specs

delete_guild_application_command(
  Remedy.Schema.User.id(),
  Remedy.Schema.Guild.id(),
  Snowflake.t()
) :: {:ok} | error()

delete_guild_emoji(guild_id, emoji_id, reason \\ nil)

Specs

delete_guild_emoji(
  Remedy.Schema.Guild.id(),
  Remedy.Schema.Emoji.id(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok}

Deletes the given emoji.

An optional reason can be provided for the audit log.

This endpoint requires the MANAGE_EMOJIS permission. It fires a t:Remedy.Consumer.guild_emojis_update/0 event.

If successful, returns {:ok}. Otherwise, returns Remedy.Api.error/0.

delete_guild_integrations(guild_id, integration_id)

Specs

delete_guild_integrations(integer(), integer()) :: error() | {:ok}

Deletes a guild integeration.

Integration to delete is specified by guild_id and integeration_id.

delete_guild_role(guild_id, role_id, reason \\ nil)

Specs

delete_guild_role(
  Remedy.Schema.Guild.id(),
  Remedy.Schema.Role.id(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok}

Deletes a role from a guild.

An optional reason can be specified for the audit log.

This endpoint requires the MANAGE_ROLES permission. It fires a t:Remedy.Consumer.guild_role_delete/0 event.

If successful, returns {:ok}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.delete_guild_role(41771983423143937, 392817238471936)

delete_interaction_followup_message(application_id \\ Cache.bot().id, token, message_id)

Specs

delete_interaction_followup_message(
  Remedy.Schema.User.id(),
  Remedy.Schema.Interaction.token(),
  Remedy.Schema.Message.id()
) :: {:ok} | error()

Delete a followup message for an interaction.

Parameters

application_id: Application ID for which to create the command. If not given, this will be fetched from Me.

token: Interaction token.

message_id: Followup message ID.

delete_invite(invite_code)

Specs

delete_invite(Invite.code()) :: error() | {:ok, Invite.simple_invite()}

Deletes an invite by its invite_code.

This endpoint requires the MANAGE_CHANNELS permission.

If successful, returns {:ok, invite}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.delete_invite("zsjUsC")

delete_message(message)

Specs

delete_message(Remedy.Schema.Message.t()) :: error() | {:ok}

Same as delete_message/2, but takes a Remedy.Struct.Message instead of a channel_id and message_id.

delete_message(channel_id, message_id)

Specs

delete_message(Remedy.Schema.Channel.id(), Remedy.Schema.Message.id()) ::
  error() | {:ok}

Deletes a message.

This endpoint requires the 'VIEW_CHANNEL' and 'MANAGE_MESSAGES' permission. It fires the MESSAGE_DELETE event.

If successful, returns {:ok}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.delete_message(43189401384091, 43189401384091)

delete_own_reaction(channel_id, message_id, emoji)

Specs

delete_own_reaction(
  Remedy.Schema.Channel.id(),
  Remedy.Schema.Message.id(),
  emoji()
) :: error() | {:ok}

Deletes a reaction the current user has made for the message.

This endpoint requires the VIEW_CHANNEL and READ_MESSAGE_HISTORY permissions. It fires a t:Remedy.Consumer.message_reaction_remove/0 event.

If successful, returns {:ok}. Otherwise, returns Remedy.Api.error/0.

See create_reaction/3 for similar examples.

delete_pinned_channel_message(channel_id, message_id)

Specs

delete_pinned_channel_message(
  Remedy.Schema.Channel.id(),
  Remedy.Schema.Message.id()
) :: error() | {:ok}

Unpins a message in a channel.

This endpoint requires the 'VIEW_CHANNEL', 'READ_MESSAGE_HISTORY', and 'MANAGE_MESSAGES' permissions. It fires the t:Remedy.Consumer.message_update/0 and t:Remedy.Consumer.channel_pins_update/0 events.

Returns {:ok} if successful. error otherwise.

delete_reaction(channel_id, message_id, emoji)

Specs

delete_reaction(Remedy.Schema.Channel.id(), Remedy.Schema.Message.id(), emoji()) ::
  error() | {:ok}

Deletes all reactions of a given emoji from a message.

This endpoint requires the MANAGE_MESSAGES permissions. It fires a t:Remedy.Consumer.message_reaction_remove_emoji/0 event.

If successful, returns {:ok}. Otherwise, returns Remedy.Api.error/0.

See create_reaction/3 for similar examples.

delete_user_reaction(channel_id, message_id, emoji, user_id)

Specs

delete_user_reaction(
  Remedy.Schema.Channel.id(),
  Remedy.Schema.Message.id(),
  emoji(),
  Remedy.Schema.User.id()
) :: error() | {:ok}

Deletes another user's reaction from a message.

This endpoint requires the VIEW_CHANNEL, READ_MESSAGE_HISTORY, and MANAGE_MESSAGES permissions. It fires a t:Remedy.Consumer.message_reaction_remove/0 event.

If successful, returns {:ok}. Otherwise, returns Remedy.Api.error/0.

See create_reaction/3 for similar examples.

delete_webhook(webhook_id, reason \\ nil)

Specs

delete_webhook(Remedy.Schema.Webhook.id(), Remedy.Schema.AuditLogEntry.reason()) ::
  error() | {:ok}

Deletes a webhook.

Parameters

webhook_id - Id of webhook to delete.

reason - An optional reason for the guild audit log.

edit_channel_permissions(channel_id, overwrite_id, permission_info, reason \\ nil)

Specs

edit_channel_permissions(
  integer(),
  integer(),
  %{
    :type => String.t(),
    optional(:allow) => integer(),
    optional(:deny) => integer()
  },
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok}

Edit the permission overwrites for a user or role.

Role or user to overwrite is specified by overwrite_id.

permission_info is a map with the following keys:

type - Required; member if editing a user, role if editing a role.

allow - Bitwise value of allowed permissions.

deny - Bitwise value of denied permissions.

type - member if editing a user, role if editing a role.

An optional reason can be provided for the audit log.

allow and deny are defaulted to 0, meaning that even if you don't specify them, they will override their respective former values in an existing overwrite.

edit_global_application_command(application_id \\ Cache.bot().id, command_id, command)

Specs

edit_global_application_command(Remedy.Schema.User.id(), Snowflake.t(), map()) ::
  {:ok, map()} | error()

Update an existing global application command.

The updated command will be available on all guilds in around an hour.

Parameters

application_id: Application ID for which to edit the command. If not given, this will be fetched from Me.

command_id: The current snowflake of the command.

command: Command configuration, see the linked API documentation for reference.

Return value

The updated command. See the official reference: https://discord.com/developers/docs/interactions/slash-commands#edit-global-application-command

edit_guild_application_command(application_id \\ Cache.bot().id, guild_id, command_id, command)

Specs

edit_guild_application_command(
  Remedy.Schema.User.id(),
  Remedy.Schema.Guild.id(),
  Snowflake.t(),
  map()
) :: {:ok, map()} | error()

edit_message(message, options)

Specs

edit_message(Remedy.Schema.Message.t(), options()) ::
  error() | {:ok, Remedy.Schema.Message.t()}

Same as edit_message/3, but takes a Remedy.Struct.Message instead of a channel_id and message_id.

edit_message(channel_id, message_id, options)

Specs

edit_message(
  Remedy.Schema.Channel.id(),
  Remedy.Schema.Message.id(),
  options() | String.t()
) :: error() | {:ok, Remedy.Schema.Message.t()}

Edits a previously sent message in a channel.

This endpoint requires the VIEW_CHANNEL permission. It fires the t:Remedy.Consumer.message_update/0 event.

If options is a string, options will be used as the message's content.

If successful, returns {:ok, message}. Otherwise, returns a Remedy.Api.error/0.

Options

:content (string) - the message contents (up to 2000 characters)

:embed (t:Remedy.Struct.Embed.t/0) - embedded rich content

Examples

iex>

Remedy.Api.edit_message(43189401384091, 1894013840914098, content: "hello world!")

Remedy.Api.edit_message(43189401384091, 1894013840914098, "hello world!")

import Remedy.Struct.Embed embed = %Remedy.Struct.Embed{} |> put_title("embed") |> put_description("new desc") Remedy.Api.edit_message(43189401384091, 1894013840914098, embed: embed)

Remedy.Api.edit_message(43189401384091, 1894013840914098, content: "hello world!", embed: embed)

execute_git_webhook(webhook_id, webhook_token, wait \\ false)

Specs

execute_git_webhook(
  Remedy.Schema.Webhook.id(),
  Remedy.Schema.Webhook.token(),
  boolean()
) :: error() | {:ok}

Executes a git webhook.

Parameters

webhook_id - Id of the webhook to execute.

webhook_token - Token of the webhook to execute.

execute_slack_webhook(webhook_id, webhook_token, wait \\ false)

Specs

execute_slack_webhook(
  Remedy.Schema.Webhook.id(),
  Remedy.Schema.Webhook.token(),
  boolean()
) :: error() | {:ok}

Executes a slack webhook.

Parameters

webhook_id - Id of the webhook to execute.

webhook_token - Token of the webhook to execute.

execute_webhook(webhook_id, webhook_token, args, wait \\ false)

Specs

execute_webhook(
  Remedy.Schema.Webhook.id() | Remedy.Schema.User.id(),
  Remedy.Schema.Webhook.token() | Remedy.Schema.Interaction.token(),
  matrix(),
  boolean()
) :: error() | {:ok}

Executes a webhook.

## Parameters

webhook_id - Id of the webhook to execute.

webhook_token - Token of the webhook to execute.

args - Map with the following required keys:

content - Message content.

file - File to send.

embeds - List of embeds to send.

username - Overrides the default name of the webhook.

avatar_url - Overrides the default avatar of the webhook.

tts - Whether the message should be read over text to speech.

wait - Whether to return an error or not. Defaults to false.

Only one of content, file or embeds should be supplied in the args parameter.

get_application_information()

Specs

get_application_information() :: error() | {:ok, map()}

Gets the bot's OAuth2 application info.

Example

iex>

Remedy.Api.get_application_information {:ok, %{ bot_public: false, bot_require_code_grant: false, description: "Test", icon: nil, id: "172150183260323840", name: "Baba O-Riley", owner: %{

avatar: nil,
discriminator: "0042",
id: "172150183260323840",
username: "i own a bot"

}, }}

get_channel(channel_id)

Specs

get_channel(Remedy.Schema.Channel.id()) ::
  error() | {:ok, Remedy.Schema.Channel.t()}

Gets a channel.

If successful, returns {:ok, channel}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex> Remedy.Api.get_channel(381889573426429952)
{:ok, %Remedy.Struct.Channel{id: 381889573426429952}}

get_channel_invites(channel_id)

Specs

get_channel_invites(Remedy.Schema.Channel.id()) ::
  error() | {:ok, [Invite.detailed_invite()]}

Gets a list of invites for a channel.

This endpoint requires the 'VIEW_CHANNEL' and 'MANAGE_CHANNELS' permissions.

If successful, returns {:ok, invite}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.getchannel_invites(43189401384091) {:ok, [%Remedy.Struct.Invite{} | ]}

get_channel_message(channel_id, message_id)

Specs

get_channel_message(Remedy.Schema.Channel.id(), Remedy.Schema.Message.id()) ::
  error() | {:ok, Remedy.Schema.Message.t()}

Retrieves a message from a channel.

This endpoint requires the 'VIEW_CHANNEL' and 'READ_MESSAGE_HISTORY' permissions.

If successful, returns {:ok, message}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.get_channel_message(43189401384091, 198238475613443)

get_channel_messages(channel_id, limit, locator \\ {})

Specs

get_channel_messages(Remedy.Schema.Channel.id(), limit(), locator()) ::
  error() | {:ok, [Remedy.Schema.Message.t()]}

Retrieves a channel's messages around a locator up to a limit.

This endpoint requires the 'VIEW_CHANNEL' permission. If the current user is missing the 'READ_MESSAGE_HISTORY' permission, then this function will return no messages.

If successful, returns {:ok, messages}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex> Remedy.Api.get_channel_messages(43189401384091, 5, {:before, 130230401384})
{:ok, %Message{}}

get_channel_webhooks(channel_id)

Specs

get_channel_webhooks(Remedy.Schema.Channel.id()) ::
  error() | {:ok, [Remedy.Struct.Webhook.t()]}

Gets a list of webook for a channel.

Parameters

channel_id - Channel to get webhooks for.

get_current_user()

Specs

get_current_user() :: error() | {:ok, Remedy.Schema.User.t()}

Gets info on the current user.

If remedy's caching is enabled, it is recommended to use Bot.get/0 instead of this function. This is because sending out an API request is much slower than pulling from our cache.

If the request is successful, this function returns {:ok, user}, where user is remedy's Remedy.Struct.User. Otherwise, returns {:error, reason}.

get_current_user_guilds(options \\ [])

Specs

get_current_user_guilds(options()) ::
  error() | {:ok, [Remedy.Schema.Guild.user_guild()]}

Gets a list of guilds the user is currently in.

This endpoint requires the guilds OAuth2 scope.

If successful, returns {:ok, guilds}. Otherwise, returns a Remedy.Api.error/0.

Options

:before (t:Sunbake.Snowflake.t/0) - get guilds before this guild ID

:after (t:Sunbake.Snowflake.t/0) - get guilds after this guild ID

:limit (integer) - max number of guilds to return (1-100)

Examples

iex>

iex> Remedy.Api.get_current_user_guilds(limit: 1) {:ok, [%Remedy.Struct.Guild{}]}

get_global_application_commands(application_id \\ Cache.bot().id)

Specs

get_global_application_commands(Remedy.Schema.User.id()) ::
  {:ok, [map()]} | error()

Fetch all global commands.

Parameters

application_id: Application ID for which to search commands. If not given, this will be fetched from Me.

Return value

A list of ApplicationCommands on success. See the official reference: https://discord.com/developers/docs/interactions/slash-commands#applicationcommand

Example

iex>

iex> Remedy.Api.get_global_application_commands {:ok, [ %{

 application_id: "455589479713865749",
 description: "ed, man! man, ed",
 id: "789841753196331029",
 name: "edit"

} ]}

get_guild(guild_id)

Specs

get_guild(Remedy.Schema.Guild.id()) ::
  error() | {:ok, Remedy.Schema.Guild.rest_guild()}

Gets a guild.

If successful, returns {:ok, guild}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.get_guild(81384788765712384) {:ok, %Remedy.Struct.Guild{id: 81384788765712384}}

get_guild_application_commands(application_id \\ Cache.bot().id, guild_id)

Specs

get_guild_application_commands(
  Remedy.Schema.User.id(),
  Remedy.Schema.Guild.id()
) :: {:ok, [map()]} | error()

get_guild_audit_log(guild_id, options \\ [])

Specs

get_guild_audit_log(Remedy.Schema.Guild.id(), options()) ::
  {:ok, Remedy.Schema.AuditLog.t()} | error()

Get the t:Remedy.Struct.Guild.AuditLog.t/0 for the given guild_id.

Options

:user_id (t:Remedy.Struct.User.id/0) - filter the log for a user ID

:action_type (integer/0) - filter the log by audit log type, see Audit Log Events

:before (t:Remedy.Struct.Snowflake.t/0) - filter the log before a certain entry ID

:limit (pos_integer/0) - how many entries are returned (default 50, minimum 1, maximum 100)

get_guild_ban(guild_id, user_id)
(since 0.5.0)

Specs

get_guild_ban(integer(), integer()) ::
  error() | {:ok, Remedy.Schema.Guild.Ban.t()}

Gets a ban object for the given user from a guild.

get_guild_bans(guild_id)

Specs

get_guild_bans(integer()) :: error() | {:ok, [Remedy.Struct.User.t()]}

Gets a list of users banned from a guild.

Guild to get bans for is specified by guild_id.

get_guild_channels(guild_id)

Specs

get_guild_channels(Remedy.Schema.Guild.id()) ::
  error() | {:ok, [Remedy.Schema.Channel.guild_channel()]}

Gets a list of guild channels.

If successful, returns {:ok, channels}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.getguild_channels(81384788765712384) {:ok, [%Remedy.Struct.Channel{guild_id: 81384788765712384} | ]}

get_guild_embed(guild_id)

Specs

get_guild_embed(integer()) :: error() | {:ok, map()}

Gets a guild embed.

get_guild_emoji(guild_id, emoji_id)

Specs

get_guild_emoji(Remedy.Schema.Guild.id(), Remedy.Schema.Emoji.id()) ::
  error() | {:ok, Remedy.Schema.Emoji.t()}

Gets an emoji for the given guild and emoji ids.

This endpoint requires the MANAGE_EMOJIS permission.

If successful, returns {:ok, emoji}. Otherwise, returns Remedy.Api.error/0.

get_guild_integrations(guild_id)

Specs

get_guild_integrations(Remedy.Schema.Guild.id()) ::
  error() | {:ok, [Remedy.Struct.Guild.Integration.t()]}

Gets a list of guild integerations.

Guild to get integrations for is specified by guild_id.

get_guild_invites(guild_id)

Specs

get_guild_invites(Remedy.Schema.Guild.id()) ::
  error() | {:ok, [Invite.detailed_invite()]}

Gets a list of invites for a guild.

This endpoint requires the MANAGE_GUILD permission.

If successful, returns {:ok, invites}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.getguild_invites(81384788765712384) {:ok, [%Remedy.Struct.Invite{} | ]}

get_guild_member(guild_id, user_id)

Specs

get_guild_member(Remedy.Schema.Guild.id(), Remedy.Schema.User.id()) ::
  error() | {:ok, Remedy.Schema.Member.t()}

Gets a guild member.

If successful, returns {:ok, member}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.get_guild_member(4019283754613, 184937267485)

get_guild_prune_count(guild_id, days)

Specs

get_guild_prune_count(Remedy.Schema.Guild.id(), 1..30) ::
  error() | {:ok, %{pruned: integer()}}

Gets the number of members that would be removed in a prune given days.

This endpoint requires the KICK_MEMBERS permission.

If successful, returns {:ok, %{pruned: pruned}}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.get_guild_prune_count(81384788765712384, 1) {:ok, %{pruned: 0}}

get_guild_roles(guild_id)

Specs

get_guild_roles(Remedy.Schema.Guild.id()) ::
  error() | {:ok, [Remedy.Schema.Role.t()]}

Gets a guild's roles.

If successful, returns {:ok, roles}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.get_guild_roles(147362948571673)

get_guild_webhooks(guild_id)

Specs

get_guild_webhooks(Remedy.Schema.Guild.id()) ::
  error() | {:ok, [Remedy.Struct.Webhook.t()]}

Gets a list of webooks for a guild.

Parameters

guild_id - Guild to get webhooks for.

get_invite(invite_code, options \\ [])

Specs

get_invite(Invite.code(), options()) :: error() | {:ok, Invite.simple_invite()}

Gets an invite by its invite_code.

If successful, returns {:ok, invite}. Otherwise, returns a Remedy.Api.error/0.

Options

:with_counts (boolean) - whether to include member count fields

Examples

iex>

Remedy.Api.get_invite("zsjUsC")

Remedy.Api.get_invite("zsjUsC", with_counts: true)

get_pinned_messages(channel_id)

Specs

get_pinned_messages(Remedy.Schema.Channel.id()) ::
  error() | {:ok, [Remedy.Schema.Message.t()]}

Retrieves all pinned messages from a channel.

This endpoint requires the 'VIEW_CHANNEL' and 'READ_MESSAGE_HISTORY' permissions.

If successful, returns {:ok, messages}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.get_pinned_messages(43189401384091)

get_reactions(channel_id, message_id, emoji)

Specs

get_reactions(Remedy.Schema.Channel.id(), Remedy.Schema.Message.id(), emoji()) ::
  error() | {:ok, [Remedy.Schema.User.t()]}

Gets all users who reacted with an emoji.

This endpoint requires the VIEW_CHANNEL and READ_MESSAGE_HISTORY permissions.

If successful, returns {:ok, users}. Otherwise, returns Remedy.Api.error/0.

See create_reaction/3 for similar examples.

get_token()

Specs

get_token() :: String.t()

Returns the token of the bot.

get_user(user_id)

Specs

get_user(Remedy.Schema.User.id()) :: error() | {:ok, Remedy.Schema.User.t()}

Gets a user by its t:Remedy.Struct.User.id/0.

If the request is successful, this function returns {:ok, user}, where user is a Remedy.Struct.User. Otherwise, returns {:error, reason}.

get_user_connections()

Specs

get_user_connections() :: error() | {:ok, list()}

Gets a list of user connections.

get_user_dms()

Specs

get_user_dms() :: error() | {:ok, [Remedy.Schema.Channel.dm_channel()]}

Gets a list of our user's DM channels.

If successful, returns {:ok, dm_channels}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.getuser_dms() {:ok, [%Remedy.Struct.Channel{type: 1} | ]}

get_voice_region(guild_id)

Specs

get_voice_region(integer()) :: error() | {:ok, [Remedy.Struct.VoiceRegion.t()]}

Gets a list of voice regions for the guild.

Guild to get voice regions for is specified by guild_id.

get_webhook(webhook_id)

Specs

get_webhook(Remedy.Schema.Webhook.id()) ::
  error() | {:ok, Remedy.Struct.Webhook.t()}

Gets a webhook by id.

Parameters

webhook_id - Id of the webhook to get.

get_webhook_with_token(webhook_id, webhook_token)

Specs

get_webhook_with_token(
  Remedy.Schema.Webhook.id(),
  Remedy.Schema.Webhook.token()
) :: error() | {:ok, Remedy.Struct.Webhook.t()}

Gets a webhook by id and token.

This method is exactly like get_webhook/1 but does not require authentication.

Parameters

webhook_id - Id of the webhook to get.

webhook_token - Token of the webhook to get.

leave_guild(guild_id)

Specs

leave_guild(integer()) :: error() | {:ok}

Leaves a guild.

Guild to leave is specified by guild_id.

list_guild_emojis(guild_id)

Specs

list_guild_emojis(Remedy.Schema.Guild.id()) ::
  error() | {:ok, [Remedy.Schema.Emoji.t()]}

Gets a list of emojis for a given guild.

This endpoint requires the MANAGE_EMOJIS permission.

If successful, returns {:ok, emojis}. Otherwise, returns Remedy.Api.error/0.

list_guild_members(guild_id, options \\ %{})

Specs

list_guild_members(Remedy.Schema.Guild.id(), options()) ::
  error() | {:ok, [Remedy.Schema.Member.t()]}

Gets a list of a guild's members.

If successful, returns {:ok, members}. Otherwise, returns a Remedy.Api.error/0.

Options

:limit (integer) - max number of members to return (1-1000) (default: 1)

:after (t:Remedy.Struct.User.id/0) - the highest user id in the previous page (default: 0)

Examples

iex>

Remedy.Api.list_guild_members(41771983423143937, limit: 1)

list_voice_regions()

Specs

list_voice_regions() :: error() | {:ok, [Remedy.Struct.VoiceRegion.t()]}

Gets a list of voice regions.

modify_channel(channel_id, options, reason \\ nil)

Specs

modify_channel(
  Remedy.Schema.Channel.id(),
  options(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok, Remedy.Schema.Channel.t()}

Modifies a channel's settings.

An optional reason can be given for the guild audit log.

If a t:Remedy.Struct.Channel.guild_channel/0 is being modified, this endpoint requires the MANAGE_CHANNEL permission. It fires a t:Remedy.Consumer.channel_update/0 event. If a t:Remedy.Struct.Channel.channel_category/0 is being modified, then this endpoint fires multiple t:Remedy.Consumer.channel_update/0 events.

If successful, returns {:ok, channel}. Otherwise, returns a Remedy.Api.error/0.

Options

:name (string) - 2-100 character channel name

:position (integer) - the position of the channel in the left-hand listing

:topic (string) (t:Remedy.Struct.Channel.text_channel/0 only) - 0-1024 character channel topic

:nsfw (boolean) (t:Remedy.Struct.Channel.text_channel/0 only) - if the channel is nsfw

:bitrate (integer) (t:Remedy.Struct.Channel.voice_channel/0 only) - the bitrate (in bits) of the voice channel; 8000 to 96000 (128000 for VIP servers)

:user_limit (integer) (t:Remedy.Struct.Channel.voice_channel/0 only) - the user limit of the voice channel; 0 refers to no limit, 1 to 99 refers to a user limit

:permission_overwrites (list of t:Remedy.Struct.Overwrite.t/0 or equivalent map) - channel or category-specific permissions

:parent_id (t:Remedy.Struct.Channel.id/0) (t:Remedy.Struct.Channel.guild_channel/0 only) - id of the new parent category for a channel

Examples

iex> Remedy.Api.modify_channel(41771983423143933, name: "elixir-remedy", topic: "remedy discussion")
{:ok, %Remedy.Struct.Channel{id: 41771983423143933, name: "elixir-remedy", topic: "remedy discussion"}}

iex> Remedy.Api.modify_channel(41771983423143933)
{:ok, %Remedy.Struct.Channel{id: 41771983423143933}}

modify_current_user(options)

Specs

modify_current_user(options()) :: error() | {:ok, Remedy.Schema.User.t()}

Changes the username or avatar of the current user.

Options

:username (string) - new username

:avatar (string) - the user's avatar as avatar data

Examples

iex>

Remedy.Api.modify_current_user(avatar: "data:image/jpeg;base64,YXl5IGJieSB1IGx1a2luIDQgc3VtIGZ1az8=")

modify_current_user_nick(guild_id, options \\ %{})

Specs

modify_current_user_nick(Remedy.Schema.Guild.id(), options()) ::
  error() | {:ok, %{nick: String.t()}}

Modifies the nickname of the current user in a guild.

If successful, returns {:ok, %{nick: nick}}. Otherwise, returns a Remedy.Api.error/0.

Options

:nick (string) - value to set users nickname to

Examples

iex>

Remedy.Api.modify_current_user_nick(41771983423143937, nick: "Remedy") {:ok, %{nick: "Remedy"}}

modify_guild(guild_id, options \\ [], reason \\ nil)

Specs

modify_guild(
  Remedy.Schema.Guild.id(),
  options(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok, Remedy.Schema.Guild.rest_guild()}

Modifies a guild's settings.

This endpoint requires the MANAGE_GUILD permission. It fires the t:Remedy.Consumer.guild_update/0 event.

An optional reason can be provided for the audit log.

If successful, returns {:ok, guild}. Otherwise, returns a Remedy.Api.error/0.

Options

:name (string) - guild name

:region (string) - guild voice region id

:verification_level (integer) - verification level

:default_message_notifications (integer) - default message notification level

:explicit_content_filter (integer) - explicit content filter level

:afk_channel_id (t:Sunbake.Snowflake.t/0) - id for afk channel

:afk_timeout (integer) - afk timeout in seconds

:icon (base64 data URI) - 128x128 jpeg image for the guild icon

:owner_id (t:Sunbake.Snowflake.t/0) - user id to transfer guild ownership to (must be owner)

:splash (base64 data URI) - 128x128 jpeg image for the guild splash (VIP only)

:system_channel_id (t:Sunbake.Snowflake.t/0) - the id of the channel to which system messages are sent

:rules_channel_id (t:Sunbake.Snowflake.t/0) - the id of the channel that is used for rules in public guilds

:public_updates_channel_id (t:Sunbake.Snowflake.t/0) - the id of the channel where admins and moderators receive notices from Discord in public guilds

Examples

iex>

Remedy.Api.modify_guild(451824027976073216, name: "Nose Drum") {:ok, %Remedy.Struct.Guild{id: 451824027976073216, name: "Nose Drum", ...}}

modify_guild_channel_positions(guild_id, positions)

Specs

modify_guild_channel_positions(Remedy.Schema.Guild.id(), [
  %{id: integer(), position: integer()}
]) :: error() | {:ok}

Reorders a guild's channels.

This endpoint requires the MANAGE_CHANNELS permission. It fires multiple t:Remedy.Consumer.channel_update/0 events.

If successful, returns {:ok, channels}. Otherwise, returns a Remedy.Api.error/0.

positions is a list of maps that each map a channel id with a position.

Examples

iex>

Remedy.Api.modify_guild_channel_positions(279093381723062272, [%{id: 351500354581692420, position: 2}])

modify_guild_embed(guild_id, options)

Specs

modify_guild_embed(integer(), map()) :: error() | {:ok, map()}

Modifies a guild embed.

modify_guild_emoji(guild_id, emoji_id, options \\ %{}, reason \\ nil)

Specs

modify_guild_emoji(
  Remedy.Schema.Guild.id(),
  Remedy.Schema.Emoji.id(),
  options(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok, Remedy.Schema.Emoji.t()}

Modify the given emoji.

This endpoint requires the MANAGE_EMOJIS permission. It fires a t:Remedy.Consumer.guild_emojis_update/0 event.

An optional reason can be provided for the audit log.

If successful, returns {:ok, emoji}. Otherwise, returns Remedy.Api.error/0.

Options

:name (string) - name of the emoji

:roles (list of t:Sunbake.Snowflake.t/0) - roles to which this emoji will be whitelisted

Examples

iex>

Remedy.Api.modify_guild_emoji(43189401384091, 4314301984301, name: "elixir", roles: [])

modify_guild_integrations(guild_id, integration_id, options)

Specs

modify_guild_integrations(integer(), integer(), %{
  expire_behaviour: integer(),
  expire_grace_period: integer(),
  enable_emoticons: boolean()
}) :: error() | {:ok}

Changes the settings and behaviours for a guild integeration.

Integration to modify is specified by guild_id and integeration_id.

options is a map with the following keys:

expire_behavior - Expiry behavior.

expire_grace_period - Period where the integration will ignore elapsed subs.

enable_emoticons - Whether emoticons should be synced.

modify_guild_member(guild_id, user_id, options \\ %{})

Specs

modify_guild_member(
  Remedy.Schema.Guild.id(),
  Remedy.Schema.User.id(),
  options()
) :: error() | {:ok}

Modifies a guild member's attributes.

This endpoint fires the t:Remedy.Consumer.guild_member_update/0 event. It situationally requires the MANAGE_NICKNAMES, MANAGE_ROLES, MUTE_MEMBERS, DEAFEN_MEMBERS, and MOVE_MEMBERS permissions.

If successful, returns {:ok}. Otherwise, returns a Remedy.Api.error/0.

Options

:nick (string) - value to set users nickname to

:roles (list of t:Sunbake.Snowflake.t/0) - array of role ids the member is assigned

:mute (boolean) - if the user is muted

:deaf (boolean) - if the user is deafened

:channel_id (t:Sunbake.Snowflake.t/0) - id of channel to move user to (if they are connected to voice)

Examples

iex>

Remedy.Api.modify_guild_member(41771983423143937, 637162356451, nick: "Remedy")

modify_guild_role(guild_id, role_id, options, reason \\ nil)

Specs

modify_guild_role(
  Remedy.Schema.Guild.id(),
  Remedy.Schema.Role.id(),
  options(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok, Remedy.Schema.Role.t()}

Modifies a guild role.

This endpoint requires the MANAGE_ROLES permission. It fires a t:Remedy.Consumer.guild_role_update/0 event.

An optional reason can be specified for the audit log.

If successful, returns {:ok, role}. Otherwise, returns a Remedy.Api.error/0.

Options

:name (string) - name of the role

:permissions (integer) - bitwise of the enabled/disabled permissions

:color (integer) - RGB color value (default: 0)

:hoist (boolean) - whether the role should be displayed separately in the sidebar

:mentionable (boolean) - whether the role should be mentionable

Examples

iex>

Remedy.Api.modify_guild_role(41771983423143937, 392817238471936, hoist: false, name: "foo-bar")

modify_guild_role_positions(guild_id, positions, reason \\ nil)

Specs

modify_guild_role_positions(
  Remedy.Schema.Guild.id(),
  [%{id: Remedy.Schema.Role.id(), position: integer()}],
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok, [Remedy.Schema.Role.t()]}

Reorders a guild's roles.

This endpoint requires the MANAGE_ROLES permission. It fires multiple t:Remedy.Consumer.guild_role_update/0 events.

If successful, returns {:ok, roles}. Otherwise, returns a Remedy.Api.error/0.

positions is a list of maps that each map a role id with a position.

Examples

iex>

Remedy.Api.modify_guild_role_positions(41771983423143937, [%{id: 41771983423143936, position: 2}])

modify_webhook(webhook_id, args, reason \\ nil)

Specs

modify_webhook(
  Remedy.Schema.Webhook.id(),
  %{name: String.t(), avatar: String.t()},
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok, Remedy.Struct.Webhook.t()}

Modifies a webhook.

Parameters

webhook_id - Id of the webhook to modify.

args - Map with the following optional keys:

name - Name of the webhook.

avatar - Base64 128x128 jpeg image for the default avatar.

reason - An optional reason for the guild audit log.

modify_webhook_with_token(webhook_id, webhook_token, args, reason \\ nil)

Specs

modify_webhook_with_token(
  Remedy.Schema.Webhook.id(),
  Remedy.Schema.Webhook.token(),
  %{name: String.t(), avatar: String.t()},
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok, Remedy.Struct.Webhook.t()}

Modifies a webhook with a token.

This method is exactly like modify_webhook/1 but does not require authentication.

Parameters

webhook_id - Id of the webhook to modify.

webhook_token - Token of the webhook to get.

args - Map with the following optional keys:

name - Name of the webhook.

avatar - Base64 128x128 jpeg image for the default avatar.

reason - An optional reason for the guild audit log.

remove_guild_ban(guild_id, user_id, reason \\ nil)

Specs

remove_guild_ban(integer(), integer(), Remedy.Schema.AuditLogEntry.reason()) ::
  error() | {:ok}

Removes a ban for a user.

User to unban is specified by guild_id and user_id. An optional reason can be specified for the audit log.

remove_guild_member(guild_id, user_id, reason \\ nil)

Specs

remove_guild_member(
  Remedy.Schema.Guild.id(),
  Remedy.Schema.User.id(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok}

Removes a member from a guild.

This event requires the KICK_MEMBERS permission. It fires a t:Remedy.Consumer.guild_member_remove/0 event.

An optional reason can be provided for the audit log with reason.

If successful, returns {:ok}. Otherwise, returns a Remedy.Api.error/0.

Examples

iex>

Remedy.Api.remove_guild_member(1453827904102291, 18739485766253)

remove_guild_member_role(guild_id, user_id, role_id, reason \\ nil)

Specs

remove_guild_member_role(
  integer(),
  integer(),
  integer(),
  Remedy.Schema.AuditLogEntry.reason()
) :: error() | {:ok}

Removes a role from a member.

Role to remove is specified by role_id. User to remove role from is specified by guild_id and user_id. An optional reason can be given for the audit log.

request(request)

request(method, route, body \\ "", options \\ [])

request_multipart(method, route, body, options \\ [])

start_typing(channel_id)

Specs

start_typing(integer()) :: error() | {:ok}

Triggers the typing indicator.

Triggers the typing indicator in the channel specified by channel_id. The typing indicator lasts for about 8 seconds and then automatically stops.

Returns {:ok} if successful. error otherwise.

sync_guild_integrations(guild_id, integration_id)

Specs

sync_guild_integrations(integer(), integer()) :: error() | {:ok}

Syncs a guild integration.

Integration to sync is specified by guild_id and integeration_id.