Remedy.Attachment (Remedy v0.6.9)

Ecto.Type implementation for Attachments

This is an intermediate type that is used to represent attachments within Remedy. It assists with uploading attachments to messages.

It allows you increased freedom when uploading attachments and images to your messages. Instead of only being able to use remote URLs, you can also upload local files, or directly input image data.

Casting

Remote URL

https://www.google.com/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png

Local File

/path/to/file.png

Image Data

 "image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASsAAACoCAMAAACPKThEAAABKVBMVEX////qQzU0qFP7vARChfT7ugA5gfTpNiQnpEr80mjZ5vz4ycWf06vqQTPqPi88gvS70fr++vnpOSntQS4zqUxAnZ5Cg/rv9f4kpEn2+f798O///vn7vwDpOzczfvREh/Tylo/tYVbvdm373dvuamD+78r++/D++Ob97ez946L+9Nn7wh71s67/1XDNtyLDVW7w+fJKsWWiwfiGrvdPjfXl7v3e8OPI2vt2wYlXtnDsWU3wgnrznJb2vbn0qqTxiYH619T94Z783Ir82Hz7yULrTT/957D83pH7xTH8zlb96bj+8sz4xrn3vp73toX2qmL1oUf1mjHkxUTMm07FZ4DIeJDOk6rSrMLWzuRZk/W638Nuv4NarKKX0KVzuqeMyKqNs/hvoPay3LyRAaJUAAAG0UlEQVR4nO2ca1fbRhBAJRtbQJAESltMRG1DbCiv1oAQmEASHuGRPpM+bUqL+f8/opJluRKWtGNpOatdz/2WHOyze8/s7OxoZUlCEARBEARBECTAwquFBdZj4IOF+vHx8UrjzcnpazRGYFVRVVXTNENR1NXt16yHk2deK6rso2qK1th+xXpIueWFIgdxddUxuKJ54srVZah1jK0oRl25toxt1uPKI1GuHFtK4xvWI8sf0a5kWVNPWQ8td8S5ckKrznpseSPWlSNrFYvTEPGuZNlo4H4YJM="

Functions

changeset(model \\ %__MODULE__{}, params)

embed_as(_)

Callback implementation for Ecto.Type.embed_as/1.

equal?(term1, term2)

Callback implementation for Ecto.Type.equal?/2.

new(path, description \\ "")