Remedy.CDN (Remedy v0.5.2)

Discord CDN interface.

Storing images and other assets can be painful, use this module to retreive assets from the discord cdn rather than storing them.

Each function takes the required parameters to directly access a resource.

Format

All images are returned as either a .png or a .gif if the asset is animated. This is done automatically.

Size

Images can be requested of a size in [16, 32, 64, 128, 256, 512, 1024, 2048, 4096]. This is given as an optional final argument to any of the functions in this module. Arguments given will be rounded to the next largest, or the largest size if you try to go over.

Functions

achievement_icon(application_id, achievement_id, icon_hash, size \\ nil)

application_asset(application_id, asset_id, size \\ nil)

application_icon(application_id, icon, size \\ nil)

appplication_cover(application_id, cover_image, size \\ nil)

custom_emoji(id, size \\ nil)

default_user_avatar(user_dicriminator, size \\ nil)

guild_banner(guild_id, guild_banner, size \\ nil)

guild_discovery_splash(guild_id, guild_discovery_splash, size \\ nil)

guild_icon(id, guild_icon, size \\ nil)

guild_splash(id, splash, size \\ nil)

sticker_pack_banner(sticker_pack_banner_asset_id, size \\ nil)

user_avatar(id, user_avatar, size \\ nil)

user_banner(user_id, user_banner, size \\ nil)