Remedy.Cache (Remedy v0.6.7)

Functions for interracting with the cache.

The cache is populated only by events received from the gateway, it is not updated from interractions with the REST api.

Types

attrs()

Specs

attrs() :: map()

changeset()

Specs

changeset() :: Ecto.Changeset.t()

reason()

Specs

reason() :: String.t()

snowflake()

Specs

snowflake() :: Snowflake.t()

Functions

app()

bot()

create_interaction(attrs)

delete_user(id)

fetch_ban(guild_id, user_id)

Specs

fetch_ban(snowflake(), snowflake()) ::
  {:error, reason()} | {:ok, Remedy.Schema.Ban.t()}

Fetch a ban from the Cache by user_id & guild_id

fetch_ban!(guild_id, user_id)

Unsafe proxy definition for fetch_ban/2.

fetch_channel(id)

Specs

fetch_channel(snowflake()) ::
  {:error, reason()} | {:ok, Remedy.Schema.Channel.t()}

Fetch a channel from the cache.

fetch_channel!(id)

Unsafe proxy definition for fetch_channel/1.

fetch_guild(id)

Fetch a guild from the cache

fetch_guild!(id)

Unsafe proxy definition for fetch_guild/1.

fetch_integration(integration_id)

Fetch an integration by ID.

fetch_role(role_id)

Specs

fetch_role(any()) :: {:error, :not_found} | {:ok, Remedy.Schema.Role.t()}

Fetch a role from the cache.

Returns {:ok, %Role{}} or {:error, reason}

fetch_role!(role_id)

Unsafe proxy definition for fetch_role/1.

fetch_user(user_id)

Fetch a user from the cache.

Returns {:ok, %User{}} or {:error, reason}

fetch_user!(id)

Unsafe proxy definition for fetch_user/1.

get_ban(guild_id, user_id)

Specs

get_ban(snowflake(), snowflake()) :: nil | Remedy.Schema.Ban.t()

get_member(guild_id, user_id)

list_bans()

Specs

list_bans() :: {:error, reason()} | {:ok, [Remedy.Schema.Ban.t()]}

List all bans

list_bans!()

Unsafe proxy definition for list_bans/0.

list_channels(guild_id \\ nil)

Specs

list_channels(snowflake()) ::
  {:error, term()} | {:ok, [Remedy.Schema.Channel.t()]}

List channels from the cache.

list_channels!(guild_id)

Unsafe proxy definition for list_channels/1.

list_guild_bans(guild_id)

Specs

list_guild_bans(snowflake()) ::
  {:error, reason()} | {:ok, [Remedy.Schema.Ban.t()]}

List all bans associated with a guild.

list_guild_bans!(guild_id)

Unsafe proxy definition for list_guild_bans/1.

list_guilds()

list_members()

List Members

list_user_bans(user_id)

Specs

list_user_bans(snowflake()) ::
  {:error, reason()} | {:ok, [Remedy.Schema.Ban.t()]}

List all bans associated with a user.

list_user_bans!(user_id)

Unsafe proxy definition for list_user_bans/1.

remove_message_reactions(message_id)

Specs

remove_message_reactions(snowflake()) ::
  {:ok, Remedy.Schema.Message.t()} | {:error, reason()}

update_guild_emojis(guild_id, params)

update_guild_stickers(guild_id, params)

update_member(attrs)

update_role!(id, attrs)

Unsafe proxy definition for update_role/2.

update_user(attrs)

user_banned?(guild_id, user_id)

Specs

user_banned?(any(), any()) :: true | false

Returns True/False to the user being banned from a guild.