Remedy.CaseHelpers (Remedy v0.6.9)

Functions within this module take heavy inspiration from the Recase library with dead weight that we do not require cut out.

Functions

to_pascal(value)

converts a string or atom to pascal case.

Examples

iex> Remedy.CaseHelpers.to_pascal("hello world")
"HelloWorld"

iex> Remedy.CaseHelpers.to_pascal(:case_helpers)
CaseHelpers

to_sentence(value)

to_title(value)

Convert a string to title case.