Remedy.Consumer behaviour (Remedy v0.5.4)

Consumer process for gateway event handling.

Consuming Gateway Events

To handle events, Remedy uses a GenStage implementation.

Remedy defines the producer and producer_consumer in the GenStage design. To consume the events you must create at least one consumer process. It is generally recommended that you spawn a consumer per core. To find this number you can use System.schedulers_online/0.

Remedy uses a ConsumerSupervisor to dispatch events, meaning your handlers will each be ran in their own seperate task.

Example

An example consumer can be found here.

Types

channel_create()

Specs

channel_create() :: {:CHANNEL_CREATE, Channel.t(), Websocket.t()}

channel_delete()

Specs

channel_delete() :: {:CHANNEL_DELETE, Channel.t(), Websocket.t()}

channel_pins_ack()

Specs

channel_pins_ack() :: {:CHANNEL_PINS_ACK, map(), Websocket.t()}

channel_pins_update()

Specs

channel_pins_update() ::
  {:CHANNEL_PINS_UPDATE, ChannelPinsUpdate.t(), Websocket.t()}

channel_update()

Specs

channel_update() :: {:CHANNEL_UPDATE, Channel.t(), Websocket.t()}

event()

Specs

event() ::
  channel_create()
  | channel_delete()
  | channel_update()
  | channel_pins_ack()
  | channel_pins_update()
  | guild_ban_add()
  | guild_ban_remove()
  | guild_create()
  | guild_available()
  | guild_unavailable()
  | guild_update()
  | guild_delete()
  | guild_emojis_update()
  | guild_integrations_update()
  | guild_member_add()
  | guild_members_chunk()
  | guild_member_remove()
  | guild_member_update()
  | guild_role_create()
  | guild_role_delete()
  | guild_role_update()
  | message_create()
  | message_delete()
  | message_delete_bulk()
  | message_update()
  | message_reaction_add()
  | message_reaction_remove()
  | message_reaction_remove_all()
  | message_ack()
  | presence_update()
  | ready()
  | resumed()
  | typing_start()
  | user_settings_update()
  | user_update()
  | voice_ready()
  | voice_speaking_update()
  | voice_state_update()
  | voice_server_update()
  | webhooks_update()

guild_available()

Specs

guild_available() :: {:GUILD_AVAILABLE, new_guild :: Guild.t(), Websocket.t()}

guild_ban_add()

Specs

guild_ban_add() :: {:GUILD_BAN_ADD, GuildBanAdd.t(), Websocket.t()}

guild_ban_remove()

Specs

guild_ban_remove() :: {:GUILD_BAN_REMOVE, GuildBanRemove.t(), Websocket.t()}

guild_create()

Specs

guild_create() :: {:GUILD_CREATE, new_guild :: Guild.t(), Websocket.t()}

guild_delete()

Specs

guild_delete() :: {:GUILD_DELETE, Guild.t(), Websocket.t()}

guild_emojis_update()

Specs

guild_emojis_update() :: {:GUILD_EMOJIS_UPDATE, [Emoji.t()], Websocket.t()}

guild_integrations_update()

Specs

guild_integrations_update() ::
  {:GUILD_INTEGRATIONS_UPDATE, GuildIntegrationsUpdate.t(), Websocket.t()}

guild_member_add()

Specs

guild_member_add() :: {:GUILD_MEMBER_ADD, Member.t(), Websocket.t()}

guild_member_remove()

Specs

guild_member_remove() ::
  {:GUILD_MEMBER_REMOVE, {guild_id :: integer(), old_member :: Member.t()},
   Websocket.t()}

guild_member_update()

Specs

guild_member_update() ::
  {:GUILD_MEMBER_UPDATE,
   {guild_id :: integer(), old_member :: Member.t() | nil,
    new_member :: Member.t()}, Websocket.t()}

guild_members_chunk()

Specs

guild_members_chunk() :: {:GUILD_MEMBERS_CHUNK, map(), Websocket.t()}

guild_role_create()

Specs

guild_role_create() :: {:GUILD_ROLE_CREATE, Guild.Role.t(), Websocket.t()}

guild_role_delete()

Specs

guild_role_delete() :: {:GUILD_ROLE_DELETE, Guild.Role.t(), Websocket.t()}

guild_role_update()

Specs

guild_role_update() :: {:GUILD_ROLE_UPDATE, Guild.Role.t(), Websocket.t()}

guild_unavailable()

Specs

guild_unavailable() :: {:GUILD_UNAVAILABLE, UnavailableGuild.t(), Websocket.t()}

guild_update()

Specs

guild_update() :: {:GUILD_UPDATE, Guild.t(), Websocket.t()}

message_ack()

Specs

message_ack() :: {:MESSAGE_ACK, map(), Websocket.t()}

message_create()

Specs

message_create() :: {:MESSAGE_CREATE, Message.t(), Websocket.t()}

message_delete()

Specs

message_delete() :: {:MESSAGE_DELETE, MessageDelete.t(), Websocket.t()}

message_delete_bulk()

Specs

message_delete_bulk() ::
  {:MESSAGE_DELETE_BULK, MessageDeleteBulk.t(), Websocket.t()}

message_reaction_add()

Specs

message_reaction_add() ::
  {:MESSAGE_REACTION_ADD, MessageReactionAdd.t(), Websocket.t()}

message_reaction_remove()

Specs

message_reaction_remove() ::
  {:MESSAGE_REACTION_REMOVE, MessageReactionRemove.t(), Websocket.t()}

message_reaction_remove_all()

Specs

message_reaction_remove_all() ::
  {:MESSAGE_REACTION_REMOVE_ALL, MessageReactionRemoveAll.t(), Websocket.t()}

message_reaction_remove_emoji()

Specs

message_reaction_remove_emoji() ::
  {:MESSAGE_REACTION_REMOVE_EMOJI, MessageReactionRemoveEmoji.t(),
   Websocket.t()}

message_update()

Specs

message_update() :: {:MESSAGE_UPDATE, Message.t(), Websocket.t()}

presence_update()

Specs

presence_update() :: {:PRESENCE_UPDATE, Presence.t(), Websocket.t()}

ready()

Specs

ready() :: {:READY, Ready.t(), Websocket.t()}

resumed()

Specs

resumed() :: {:RESUMED, map(), Websocket.t()}

typing_start()

Specs

typing_start() :: {:TYPING_START, TypingStart.t(), Websocket.t()}

user_settings_update()

Specs

user_settings_update() :: no_return()

user_update()

Specs

user_update() ::
  {:USER_UPDATE, {old_user :: User.t() | nil, new_user :: User.t()},
   Websocket.t()}

voice_ready()

Specs

voice_ready() :: {:VOICE_READY, VoiceReady.t(), VoiceWSState.t()}

voice_server_update()

Specs

voice_server_update() ::
  {:VOICE_SERVER_UPDATE, VoiceServerUpdate.t(), Websocket.t()}

voice_speaking_update()

Specs

voice_speaking_update() ::
  {:VOICE_SPEAKING_UPDATE, SpeakingUpdate.t(), VoiceWSState.t()}

voice_state_update()

Specs

voice_state_update() :: {:VOICE_STATE_UPDATE, VoiceState.t(), Websocket.t()}

webhooks_update()

Specs

webhooks_update() :: {:WEBHOOKS_UPDATE, map(), Websocket.t()}

Callbacks

handle_event(event)

Specs

handle_event(event()) :: any()

Functions

start_link(mod, opts \\ [])