Remedy.Embed (Remedy v0.6.9)

Convenience functions for working with embeds.

Using the helpers in this module you can coneniently convert various objects into fields within an embed, allowing easier manipulation and more consistent embeds with less boilerplate.

Consider you wish to create an embed regarding an action that a user took. We can use pre established objects to populate the embed fields. For example:

user = %User{id: 1, name: "John Doe"}
message = %Message{text: "Hello World", timestamp: timestamp}

%Embed{}
|> put_author(user)
|> put_timestamp(timestamp)
|> put_colour("#F34AC3")
|> put_title("User Silenced")
|> put_description("User was silenced for breaking the rules.")

Functions

append_field(embed, name, value, inline \\ true)

append a field to the embed.

prepend_field(embed, name, value, inline \\ true)

append a field to the embed.

put_author(embed)

Specs

put_author(Remedy.Schema.Embed.t()) :: Remedy.Schema.Embed.t()

Add the bot as the author.

put_author(embed, member)

Specs

put_author(Remedy.Schema.Embed.t(), Remedy.Schema.Member.t()) ::
  Remedy.Schema.Embed.t()

put_author(Remedy.Schema.Embed.t(), Remedy.Schema.User.t()) ::
  Remedy.Schema.Embed.t()

Update the Author to a user or member.

put_colour(embed, color)

Specs

put_colour(Remedy.Schema.Embed.t(), Remedy.Colour.c()) ::
  Remedy.Schema.Embed.t()

Update the embed colour.

put_description(embed, description)

Specs

put_description(Remedy.Schema.Embed.t(), String.t()) :: Remedy.Schema.Embed.t()

Update the embed description.

put_timestamp(embed, timestamp)

Update the embeds timestamp.