Remedy.Gateway.Dispatch.VoiceStateUpdate (Remedy v0.6.7)

Voice State Update Event

Types

channel_id()

Specs

channel_id() :: Sunbake.Snowflake.t()

Channel ID this voice state is for

deaf?()

Specs

deaf?() :: boolean()

Whether this user is deafened by the server

guild_id()

Specs

guild_id() :: Sunbake.Snowflake.t()

Guild ID this voice state is for, if applicable

member()

Specs

member() :: Remedy.Schema.Member.t() | nil

Guild member this voice state is for, if applicable

mute?()

Specs

mute?() :: boolean()

Whether this user is muteened by the server

request_to_speak_timestamp()

Specs

request_to_speak_timestamp() :: DateTime.t() | nil

Time at which the user requested to speak, if applicable

self_deaf?()

Specs

self_deaf?() :: boolean()

Whether this user is locally deafened

self_mute?()

Specs

self_mute?() :: boolean()

Whether this user is locally muted

self_stream?()

Specs

self_stream?() :: boolean()

Whether the user is streaming using "Go Live"

self_video?()

Specs

self_video?() :: boolean()

Whether this user's camera is enabled

session_id()

Specs

session_id() :: String.t()

Session ID for this voice state

suppress?()

Specs

suppress?() :: boolean()

Whether this user is muted by the current user

t()

Specs

t() :: %Remedy.Gateway.Dispatch.VoiceStateUpdate{
  channel_id: channel_id(),
  deaf?: deaf?(),
  guild_id: guild_id(),
  member: member(),
  mute?: mute?(),
  request_to_speak_timestamp: request_to_speak_timestamp(),
  self_deaf?: self_deaf?(),
  self_mute?: self_mute?(),
  self_stream?: self_stream?(),
  self_video?: self_video?(),
  session_id: session_id(),
  suppress?: suppress?(),
  user_id: user_id()
}

user_id()

Specs

user_id() :: Sunbake.Snowflake.t()

User this voice state is for

Functions

changeset(model, params)

handle(arg)