Remedy.Gateway.Intents (Remedy v0.6.9)

Ecto.Type compatible bit flag type.

Bit	Flag

1 << 0	:GUILDS

1 << 1	:GUILD_MEMBERS

1 << 2	:GUILD_BANS

1 << 3	:GUILD_EMOJIS_AND_STICKERS

1 << 4	:GUILD_INTEGRATIONS

1 << 5	:GUILD_WEBHOOKS

1 << 6	:GUILD_INVITES

1 << 7	:GUILD_VOICE_STATES

1 << 8	:GUILD_PRESENCES

1 << 9	:GUILD_MESSAGES

1 << 10	:GUILD_MESSAGE_REACTIONS

1 << 11	:GUILD_MESSAGE_TYPING

1 << 12	:DIRECT_MESSAGES

1 << 13	:DIRECT_MESSAGE_REACTIONS

1 << 14	:DIRECT_MESSAGE_TYPING

1 << 15	:GUILD_SCHEDULED_EVENTS

Casting

The following are examples of valid inputs for casting. Regardless of the format provided, values will be cast to an integer/0 value for storage.

Decimal Integer

14092

Binary Integer

0b100000000

String List

["DIRECT_MESSAGE_TYPING", "GUILDS", "GUILD_BANS", "GUILD_INTEGRATIONS",
 "GUILD_INVITES", "GUILD_MESSAGES", "GUILD_MESSAGE_REACTIONS",
 "GUILD_PRESENCES", "GUILD_VOICE_STATES"]

Atom List

[:DIRECT_MESSAGE_TYPING, :GUILDS, :GUILD_BANS, :GUILD_WEBHOOKS]

Boolean Map

%{
  DIRECT_MESSAGES: true,
  DIRECT_MESSAGE_TYPING: true,
  GUILDS: true,
  GUILD_MESSAGE_REACTIONS: true,
  GUILD_MESSAGE_TYPING: true
}

Types

c()

Specs

c() :: t() | [f()] | map() | [String.t()]

Castable to Intents Type

f()

Specs

f() ::
  :DIRECT_MESSAGES
  | :DIRECT_MESSAGE_REACTIONS
  | :DIRECT_MESSAGE_TYPING
  | :GUILDS
  | :GUILD_BANS
  | :GUILD_EMOJIS_AND_STICKERS
  | :GUILD_INTEGRATIONS
  | :GUILD_INVITES
  | :GUILD_MEMBERS
  | :GUILD_MESSAGES
  | :GUILD_MESSAGE_REACTIONS
  | :GUILD_MESSAGE_TYPING
  | :GUILD_PRESENCES
  | :GUILD_SCHEDULED_EVENTS
  | :GUILD_VOICE_STATES
  | :GUILD_WEBHOOKS

Intents Flags

load()

Specs

load() :: [f()]

Loaded Intents Flags

t()

Specs

t() :: 0..65535

Intents Type