Remedy.Gateway.Payload behaviour (Remedy v0.5.3)

Payload represents the data packet set to discord through the API. All processing of the payload should be done within this context as it will be cleared upon returning to the session process.

These functions exist to attempt friendly discourse with the ill mannered Discord gateway. Documentation is included for completeness but using them is at your peril!

Fields

op: Opcode.

d: Data.

s: Sequence.

t: Event Name.

Types

opts()

Specs

opts() :: list() | nil

payload()

Specs

payload() :: any() | nil

socket()

Specs

socket() :: Websocket.t()

Callbacks

digest(socket, payload)
(optional)

Specs

digest(socket(), payload()) :: socket()

Digest the data frame from Discord and loads the data into the socket. For example:

The :heartbeat_ack flag on the websocket needs to be set to true once the :HEARTBEAT_ACK event is received from discord.

In short. Do what you need to do with the payload. because its going away

payload(socket, opts)
(optional)

Specs

payload(socket(), opts()) :: {payload(), socket()}

Describes how to take the current socket state and construct the payload data for this modules event type.

For example: Heartbeat.send/2 will take the socket, and a keyword list of options, and construct the payload data for the event of type :HEARTBEAT. It is the responsibility of the developer to ensure that all events required to be sent implement this function.

If the behaviour is not described. Passing this function will just pass the socket back to session to continue doing what it do.

Functions

digest(socket, event, payload)

Specs

digest(Remedy.Gateway.Websocket.t(), any(), binary()) ::
  Remedy.Gateway.Websocket.t()

Digest the data frame into the socket.

Delegates the digestion to the appropriate module. It is that modules responsibility to implement the digest/2 callback and return the %Websocket{}

send(socket, event, opts \\ [])

Specs

send(Remedy.Gateway.Websocket.t(), any(), any()) :: any()

Sends a command based on the current websocket state.

Using this will invoke the appropriate modules