Remedy.Gateway.WSState (Remedy v0.6.7)

Contains all the information required to maintain the gateway websocket connection to Discord.

This is provided to allow the user to enact custom logic on the gateway events as they are received from the consumer. It should be noted that the websocket state consumed with an event is a 'snapshot' of the state at the time that event was received, and does not relate to the current state of the websocket.

Types

conn()

Specs

conn() :: pid()

gateway()

Specs

gateway() :: String.t()

gun_conn()

Specs

gun_conn() :: pid()

gun_data_stream()

Specs

gun_data_stream() :: reference()

heartbeat_ack()

Specs

heartbeat_ack() :: boolean()

heartbeat_interval()

Specs

heartbeat_interval() :: integer()

heartbeat_timer()

Specs

heartbeat_timer() :: integer()

last_heartbeat_ack()

Specs

last_heartbeat_ack() :: DateTime.t() | nil

last_heartbeat_send()

Specs

last_heartbeat_send() :: DateTime.t() | nil

payload_dispatch_event()

Specs

payload_dispatch_event() :: atom()

payload_op_code()

Specs

payload_op_code() :: integer()

payload_op_event()

Specs

payload_op_event() :: atom()

payload_sequence()

Specs

payload_sequence() :: integer()

session_id()

Specs

session_id() :: String.t() | nil

shard()

Specs

shard() :: integer()

shard_pid()

Specs

shard_pid() :: pid()

t()

Specs

t() :: %Remedy.Gateway.WSState{
  conn: conn(),
  gateway: gateway(),
  gun_conn: gun_conn(),
  gun_data_stream: gun_data_stream(),
  heartbeat_ack: heartbeat_ack(),
  heartbeat_interval: heartbeat_interval(),
  heartbeat_timer: heartbeat_timer(),
  last_heartbeat_ack: last_heartbeat_ack(),
  last_heartbeat_send: last_heartbeat_send(),
  payload_dispatch_event: payload_dispatch_event(),
  payload_op_code: payload_op_code(),
  payload_op_event: payload_op_event(),
  payload_sequence: payload_sequence(),
  session_id: session_id(),
  shard: shard(),
  shard_pid: term(),
  token: token(),
  v: term(),
  zlib_context: zlib_context()
}

token()

Specs

token() :: String.t()

zlib_context()

Specs

zlib_context() :: term()

Functions

get_shard_latency(state)

Gets the latency of the shard connection from a Remedy.Struct.WSState.t() struct.

Returns the latency in milliseconds as an integer, returning nil if unknown.