Remedy.Schema.AuditLogActionType (Remedy v0.6.9)

Ecto.Type compatible Audit Log Action Type.

Value	Type

1	:GUILD_UPDATE

10	:CHANNEL_CREATE

11	:CHANNEL_UPDATE

12	:CHANNEL_DELETE

13	:CHANNEL_OVERWRITE_CREATE

14	:CHANNEL_OVERWRITE_UPDATE

15	:CHANNEL_OVERWRITE_DELETE

20	:MEMBER_KICK

21	:MEMBER_PRUNE

22	:MEMBER_BAN_ADD

23	:MEMBER_BAN_REMOVE

24	:MEMBER_UPDATE

25	:MEMBER_ROLE_UPDATE

26	:MEMBER_MOVE

27	:MEMBER_DISCONNECT

28	:BOT_ADD

30	:ROLE_CREATE

31	:ROLE_UPDATE

32	:ROLE_DELETE

40	:INVITE_CREATE

41	:INVITE_UPDATE

42	:INVITE_DELETE

50	:WEBHOOK_CREATE

51	:WEBHOOK_UPDATE

52	:WEBHOOK_DELETE

60	:EMOJI_CREATE

61	:EMOJI_UPDATE

62	:EMOJI_DELETE

72	:MESSAGE_DELETE

73	:MESSAGE_BULK_DELETE

74	:MESSAGE_PIN

75	:MESSAGE_UNPIN

80	:INTEGRATION_CREATE

81	:INTEGRATION_UPDATE

82	:INTEGRATION_DELETE

83	:STAGE_INSTANCE_CREATE

84	:STAGE_INSTANCE_UPDATE

85	:STAGE_INSTANCE_DELETE

90	:STICKER_CREATE

91	:STICKER_UPDATE

92	:STICKER_DELETE

100	:GUILD_SCHEDULED_EVENT_CREATE

101	:GUILD_SCHEDULED_EVENT_UPDATE

102	:GUILD_SCHEDULED_EVENT_DELETE

110	:THREAD_CREATE

111	:THREAD_UPDATE

112	:THREAD_DELETE

Casting

The following are examples of valid inputs for casting. Regardless of the format provided, values will be cast to an integer/0 value for storage.

Integer Type

91

String Type

"EMOJI_DELETE"

Atom Type

:WEBHOOK_DELETE

Types

c()

Specs

c() :: t() | f() | String.t()

Castable to Audit Log Action Type Type

See Casting for more information.

f()

Specs

f() ::
  :CHANNEL_OVERWRITE_DELETE
  | :MEMBER_ROLE_UPDATE
  | :THREAD_DELETE
  | :CHANNEL_CREATE
  | :MESSAGE_DELETE
  | :MEMBER_MOVE
  | :CHANNEL_UPDATE
  | :MEMBER_BAN_ADD
  | :STAGE_INSTANCE_DELETE
  | :INVITE_CREATE
  | :ROLE_CREATE
  | :GUILD_SCHEDULED_EVENT_UPDATE
  | :MEMBER_BAN_REMOVE
  | :MEMBER_UPDATE
  | :INVITE_UPDATE
  | :STICKER_UPDATE
  | :STAGE_INSTANCE_UPDATE
  | :STICKER_DELETE
  | :ROLE_UPDATE
  | :ROLE_DELETE
  | :EMOJI_CREATE
  | :INTEGRATION_DELETE
  | :THREAD_CREATE
  | :CHANNEL_OVERWRITE_CREATE
  | :INVITE_DELETE
  | :EMOJI_UPDATE
  | :MEMBER_DISCONNECT
  | :WEBHOOK_CREATE
  | :WEBHOOK_DELETE
  | :MESSAGE_PIN
  | :GUILD_SCHEDULED_EVENT_CREATE
  | :MESSAGE_BULK_DELETE
  | :WEBHOOK_UPDATE
  | :MEMBER_PRUNE
  | :STAGE_INSTANCE_CREATE
  | :BOT_ADD
  | :STICKER_CREATE
  | :MEMBER_KICK
  | :EMOJI_DELETE
  | :INTEGRATION_UPDATE
  | :GUILD_UPDATE
  | :CHANNEL_DELETE
  | :GUILD_SCHEDULED_EVENT_DELETE
  | :INTEGRATION_CREATE
  | :CHANNEL_OVERWRITE_UPDATE
  | :MESSAGE_UNPIN
  | :THREAD_UPDATE

Audit Log Action Type Values

load()

Specs

load() :: f()

Loaded Audit Log Action Type

t()

Specs

t() :: 0..112

Audit Log Action Type Type