Remedy.Schema.CallbackDataFlags (Remedy v0.6.9)

Ecto.Type compatible bit flag type.

Bit	Flag

1 << 6	:EPHEMERAL

Casting

The following are examples of valid inputs for casting. Regardless of the format provided, values will be cast to an integer/0 value for storage.

Decimal Integer

64

Binary Integer

0b1000000

String List

["EPHEMERAL"]

Atom List

[:EPHEMERAL]

Boolean Map

%{EPHEMERAL: true}

Types

c()

Specs

c() :: t() | [f()] | map() | [String.t()]

Castable to Callback Data Flags Type

f()

Specs

f() :: :EPHEMERAL

Callback Data Flags Flags

load()

Specs

load() :: [f()]

Loaded Callback Data Flags Flags

t()

Specs

t() :: 0..64

Callback Data Flags Type