Remedy.Schema.CommandType (Remedy v0.6.9)

Ecto.Type compatible Command Type.

Value	Type

1	:CHAT_INPUT

2	:USER

3	:MESSAGE

Casting

The following are examples of valid inputs for casting. Regardless of the format provided, values will be cast to an integer/0 value for storage.

Integer Type

3

String Type

"MESSAGE"

Atom Type

:USER

Types

c()

Specs

c() :: t() | f() | String.t()

Castable to Command Type Type

See Casting for more information.

f()

Specs

f() :: :CHAT_INPUT | :MESSAGE | :USER

Command Type Values

load()

Specs

load() :: f()

Loaded Command Type

t()

Specs

t() :: 0..3

Command Type Type