Remedy.Schema.GuildFeatures (Remedy v0.6.9)

Ecto.Type compatible bit flag type.

Bit	Flag

1 << 0	:ANIMATED_ICON

1 << 1	:BANNER

1 << 2	:COMMERCE

1 << 3	:COMMUNITY

1 << 4	:DISCOVERABLE

1 << 5	:FEATURABLE

1 << 6	:INVITE_SPLASH

1 << 7	:MEMBER_VERIFICATION_GATE_ENABLED

1 << 8	:MONETIZATION_ENABLED

1 << 9	:MORE_STICKERS

1 << 10	:NEWS

1 << 11	:PARTNERED

1 << 12	:PREVIEW_ENABLED

1 << 13	:PRIVATE_THREADS

1 << 14	:ROLE_ICONS

1 << 15	:SEVEN_DAY_THREAD_ARCHIVE

1 << 16	:THREE_DAY_THREAD_ARCHIVE

1 << 17	:VANITY_URL

1 << 18	:VERIFIED

1 << 19	:VIP_REGIONS

1 << 20	:WELCOME_SCREEN_ENABLED

1 << 21	:NEW_THREAD_PERMISSIONS

1 << 22	:MEMBER_PROFILES

1 << 23	:THREADS_ENABLED

1 << 24	:CHANNEL_BANNER

1 << 25	:ANIMATED_BANNER

Casting

The following are examples of valid inputs for casting. Regardless of the format provided, values will be cast to an integer/0 value for storage.

Decimal Integer

66097092

Binary Integer

0b10011011010001101101010001

String List

["DISCOVERABLE", "FEATURABLE", "MORE_STICKERS", "ROLE_ICONS"]

Atom List

[:ANIMATED_ICON, :CHANNEL_BANNER, :COMMERCE, :MEMBER_PROFILES, :VANITY_URL,
 :WELCOME_SCREEN_ENABLED]

Boolean Map

%{DISCOVERABLE: true}

Types

c()

Specs

c() :: t() | [f()] | map() | [String.t()]

Castable to Guild Features Type

f()

Specs

f() ::
  :ANIMATED_BANNER
  | :ANIMATED_ICON
  | :BANNER
  | :CHANNEL_BANNER
  | :COMMERCE
  | :COMMUNITY
  | :DISCOVERABLE
  | :FEATURABLE
  | :INVITE_SPLASH
  | :MEMBER_PROFILES
  | :MEMBER_VERIFICATION_GATE_ENABLED
  | :MONETIZATION_ENABLED
  | :MORE_STICKERS
  | :NEWS
  | :NEW_THREAD_PERMISSIONS
  | :PARTNERED
  | :PREVIEW_ENABLED
  | :PRIVATE_THREADS
  | :ROLE_ICONS
  | :SEVEN_DAY_THREAD_ARCHIVE
  | :THREADS_ENABLED
  | :THREE_DAY_THREAD_ARCHIVE
  | :VANITY_URL
  | :VERIFIED
  | :VIP_REGIONS
  | :WELCOME_SCREEN_ENABLED

Guild Features Flags

load()

Specs

load() :: [f()]

Loaded Guild Features Flags

t()

Specs

t() :: 0..67_108_863

Guild Features Type