Remedy.Schema.MessageType (Remedy v0.6.9)

Ecto.Type compatible Message Type.

Value	Type

0	:DEFAULT

1	:RECIPIENT_ADD

2	:RECIPIENT_REMOVE

3	:CALL

4	:CHANNEL_NAME_CHANGE

5	:CHANNEL_ICON_CHANGE

6	:CHANNEL_PINNED_MESSAGE

7	:GUILD_MEMBER_JOIN

8	:USER_PREMIUM_GUILD_SUBSCRIPTION

9	:USER_PREMIUM_GUILD_SUBSCRIPTION_TIER_1

10	:USER_PREMIUM_GUILD_SUBSCRIPTION_TIER_2

11	:USER_PREMIUM_GUILD_SUBSCRIPTION_TIER_3

12	:CHANNEL_FOLLOW_ADD

14	:GUILD_DISCOVERY_DISQUALIFIED

15	:GUILD_DISCOVERY_REQUALIFIED

16	:GUILD_DISCOVERY_GRACE_PERIOD_INITIAL_WARNING

17	:GUILD_DISCOVERY_GRACE_PERIOD_FINAL_WARNING

18	:THREAD_CREATED

19	:REPLY

20	:CHAT_INPUT_COMMAND

21	:THREAD_STARTER_MESSAGE

22	:GUILD_INVITE_REMINDER

23	:CONTEXT_MENU_COMMAND

Casting

The following are examples of valid inputs for casting. Regardless of the format provided, values will be cast to an integer/0 value for storage.

Integer Type

8

String Type

"USER_PREMIUM_GUILD_SUBSCRIPTION_TIER_3"

Atom Type

:CHANNEL_ICON_CHANGE

Types

c()

Specs

c() :: t() | f() | String.t()

Castable to Message Type Type

See Casting for more information.

f()

Specs

f() ::
  :CALL
  | :CHANNEL_FOLLOW_ADD
  | :CHANNEL_ICON_CHANGE
  | :CHANNEL_NAME_CHANGE
  | :CHANNEL_PINNED_MESSAGE
  | :CHAT_INPUT_COMMAND
  | :CONTEXT_MENU_COMMAND
  | :DEFAULT
  | :GUILD_DISCOVERY_DISQUALIFIED
  | :GUILD_DISCOVERY_GRACE_PERIOD_FINAL_WARNING
  | :GUILD_DISCOVERY_GRACE_PERIOD_INITIAL_WARNING
  | :GUILD_DISCOVERY_REQUALIFIED
  | :GUILD_INVITE_REMINDER
  | :GUILD_MEMBER_JOIN
  | :RECIPIENT_ADD
  | :RECIPIENT_REMOVE
  | :REPLY
  | :THREAD_CREATED
  | :THREAD_STARTER_MESSAGE
  | :USER_PREMIUM_GUILD_SUBSCRIPTION
  | :USER_PREMIUM_GUILD_SUBSCRIPTION_TIER_1
  | :USER_PREMIUM_GUILD_SUBSCRIPTION_TIER_2
  | :USER_PREMIUM_GUILD_SUBSCRIPTION_TIER_3

Message Type Values

load()

Specs

load() :: f()

Loaded Message Type

t()

Specs

t() :: 0..23

Message Type Type