Remedy.Schema.Permission (Remedy v0.6.9)

Ecto.Type compatible bit flag type.

Bit	Flag

1 << 0	:CREATE_INSTANT_INVITE

1 << 1	:KICK_MEMBERS

1 << 2	:BAN_MEMBERS

1 << 3	:ADMINISTRATOR

1 << 4	:MANAGE_CHANNELS

1 << 5	:MANAGE_GUILD

1 << 6	:ADD_REACTIONS

1 << 7	:VIEW_AUDIT_LOG

1 << 8	:PRIORITY_SPEAKER

1 << 9	:STREAM

1 << 10	:VIEW_CHANNEL

1 << 11	:SEND_MESSAGES

1 << 12	:SEND_TTS_MESSAGES

1 << 13	:MANAGE_MESSAGES

1 << 14	:EMBED_LINKS

1 << 15	:ATTACH_FILES

1 << 16	:READ_MESSAGE_HISTORY

1 << 17	:MENTION_EVERYONE

1 << 18	:USE_EXTERNAL_EMOJIS

1 << 19	:VIEW_GUILD_INSIGHTS

1 << 20	:CONNECT

1 << 21	:SPEAK

1 << 22	:MUTE_MEMBERS

1 << 23	:DEAFEN_MEMBERS

1 << 24	:MOVE_MEMBERS

1 << 25	:USE_VAD

1 << 26	:CHANGE_NICKNAME

1 << 27	:MANAGE_NICKNAMES

1 << 28	:MANAGE_ROLES

1 << 29	:MANAGE_WEBHOOKS

1 << 30	:MANAGE_EMOJIS_AND_STICKERS

1 << 31	:USE_APPLICATION_COMMANDS

1 << 32	:REQUEST_TO_SPEAK

1 << 34	:MANAGE_THREADS

1 << 35	:USE_PUBLIC_THREADS

1 << 36	:USE_PRIVATE_THREADS

1 << 37	:USE_EXTERNAL_STICKERS

Casting

The following are examples of valid inputs for casting. Regardless of the format provided, values will be cast to an integer/0 value for storage.

Decimal Integer

23046934068

Binary Integer

0b100110010010000001100100000000000000

String List

["ADD_REACTIONS", "ATTACH_FILES", "CHANGE_NICKNAME", "KICK_MEMBERS",
 "MANAGE_GUILD", "MANAGE_ROLES", "MANAGE_WEBHOOKS", "MOVE_MEMBERS",
 "SEND_MESSAGES", "SPEAK", "USE_PRIVATE_THREADS", "VIEW_GUILD_INSIGHTS"]

Atom List

[:BAN_MEMBERS, :EMBED_LINKS, :MANAGE_GUILD, :MANAGE_MESSAGES, :MANAGE_ROLES,
 :MANAGE_THREADS, :MENTION_EVERYONE, :READ_MESSAGE_HISTORY, :STREAM, :USE_VAD]

Boolean Map

%{ATTACH_FILES: true, CHANGE_NICKNAME: true, CONNECT: true}

Types

c()

Specs

c() :: t() | [f()] | map() | [String.t()]

Castable to Permission Type

f()

Specs

f() ::
  :USE_PUBLIC_THREADS
  | :VIEW_CHANNEL
  | :MANAGE_WEBHOOKS
  | :ATTACH_FILES
  | :MANAGE_CHANNELS
  | :VIEW_AUDIT_LOG
  | :MANAGE_GUILD
  | :CONNECT
  | :PRIORITY_SPEAKER
  | :USE_VAD
  | :SEND_TTS_MESSAGES
  | :KICK_MEMBERS
  | :ADMINISTRATOR
  | :STREAM
  | :VIEW_GUILD_INSIGHTS
  | :USE_APPLICATION_COMMANDS
  | :MANAGE_THREADS
  | :MUTE_MEMBERS
  | :USE_EXTERNAL_STICKERS
  | :REQUEST_TO_SPEAK
  | :SEND_MESSAGES
  | :MANAGE_NICKNAMES
  | :EMBED_LINKS
  | :DEAFEN_MEMBERS
  | :MOVE_MEMBERS
  | :MENTION_EVERYONE
  | :USE_EXTERNAL_EMOJIS
  | :CREATE_INSTANT_INVITE
  | :BAN_MEMBERS
  | :ADD_REACTIONS
  | :MANAGE_ROLES
  | :USE_PRIVATE_THREADS
  | :SPEAK
  | :READ_MESSAGE_HISTORY
  | :MANAGE_EMOJIS_AND_STICKERS
  | :CHANGE_NICKNAME
  | :MANAGE_MESSAGES

Permission Flags

load()

Specs

load() :: [f()]

Loaded Permission Flags

t()

Specs

t() :: 0..266_287_972_351

Permission Type