Remedy.Schema.ResponseType (Remedy v0.6.9)

Ecto.Type compatible Response Type.

Value	Type

1	:PONG

4	:CHANNEL_MESSAGE_WITH_SOURCE

5	:DEFERRED_CHANNEL_MESSAGE_WITH_SOURCE

6	:DEFERRED_UPDATE_MESSAGE

7	:APPLICATION_UPDATE

8	:APPLICATION_COMMAND_AUTOCOMPLETE_RESULT

Casting

The following are examples of valid inputs for casting. Regardless of the format provided, values will be cast to an integer/0 value for storage.

Integer Type

6

String Type

"PONG"

Atom Type

:DEFERRED_CHANNEL_MESSAGE_WITH_SOURCE

Types

c()

Specs

c() :: t() | f() | String.t()

Castable to Response Type Type

See Casting for more information.

f()

Specs

f() ::
  :APPLICATION_COMMAND_AUTOCOMPLETE_RESULT
  | :APPLICATION_UPDATE
  | :CHANNEL_MESSAGE_WITH_SOURCE
  | :DEFERRED_CHANNEL_MESSAGE_WITH_SOURCE
  | :DEFERRED_UPDATE_MESSAGE
  | :PONG

Response Type Values

load()

Specs

load() :: f()

Loaded Response Type

t()

Specs

t() :: 0..8

Response Type Type