Remedy.Schema.UserFlags (Remedy v0.6.9)

Ecto.Type compatible bit flag type.

Bit	Flag

1 << 0	:DISCORD_EMPLOYEE

1 << 1	:PARTNERED_SERVER_OWNER

1 << 2	:HYPESQUAD_EVENTS

1 << 3	:BUG_HUNTER_LEVEL_1

1 << 6	:HYPESQUAD_BRAVERY

1 << 7	:HYPESQUAD_BRILLIANCE

1 << 8	:HYPESQUAD_BALANCE

1 << 9	:EARLY_NITRO_SUPPORTER

1 << 10	:TEAM_PSEUDO_USER

1 << 12	:SYSTEM

1 << 14	:BUG_HUNTER_LEVEL_2

1 << 16	:VERIFIED_BOT

1 << 17	:VERIFIED_DEVELOPER

1 << 18	:CERTIFIED_MODERATOR

1 << 19	:BOT_HTTP_INTERACTIONS

1 << 20	:SPAMMER

Casting

The following are examples of valid inputs for casting. Regardless of the format provided, values will be cast to an integer/0 value for storage.

Decimal Integer

725129

Binary Integer

0b1000000000

String List

["BOT_HTTP_INTERACTIONS", "BUG_HUNTER_LEVEL_1", "BUG_HUNTER_LEVEL_2",
 "EARLY_NITRO_SUPPORTER", "HYPESQUAD_BALANCE", "HYPESQUAD_BRAVERY",
 "HYPESQUAD_BRILLIANCE", "PARTNERED_SERVER_OWNER", "VERIFIED_DEVELOPER"]

Atom List

[:SYSTEM]

Boolean Map

%{DISCORD_EMPLOYEE: true, HYPESQUAD_BRAVERY: true, VERIFIED_DEVELOPER: true}

Types

c()

Specs

c() :: t() | [f()] | map() | [String.t()]

Castable to User Flags Type

f()

Specs

f() ::
  :BOT_HTTP_INTERACTIONS
  | :BUG_HUNTER_LEVEL_1
  | :BUG_HUNTER_LEVEL_2
  | :CERTIFIED_MODERATOR
  | :DISCORD_EMPLOYEE
  | :EARLY_NITRO_SUPPORTER
  | :HYPESQUAD_BALANCE
  | :HYPESQUAD_BRAVERY
  | :HYPESQUAD_BRILLIANCE
  | :HYPESQUAD_EVENTS
  | :PARTNERED_SERVER_OWNER
  | :SPAMMER
  | :SYSTEM
  | :TEAM_PSEUDO_USER
  | :VERIFIED_BOT
  | :VERIFIED_DEVELOPER

User Flags Flags

load()

Specs

load() :: [f()]

Loaded User Flags Flags

t()

Specs

t() :: 0..2_054_095

User Flags Type