Remedy.TimeHelpers (Remedy v0.6.9)

Functions and Guards for working with time.

There are four formats of concern in this module:

Snowflake - a 64-bit integer representing the number of milliseconds since the discord epoch + metadata.

t:DateTime/0 - a datetime.datetime object.

Remedy.ISO8601.t/0 - a string in the ISO8601 format.

UnixTime - a number representing the number of seconds since the unix epoch.

Guards

is_datetime(value)
(macro)

Guard to test if a value is a DateTime.

Functions

discord_epoch()

Specs

discord_epoch() :: 1_420_070_400_000

Returns the discord epoch.

is_datetime?(value)

Specs

is_datetime?(any()) :: boolean()

Check if the term is a valid DateTime.t/0.

Examples

iex> date_time = DateTime.now!("Etc/UTC")
...> Remedy.TimeHelpers.is_datetime?(date_time)
true

iex> "2020-01-01T00:00:00Z" |> Remedy.TimeHelpers.is_datetime?()
false

is_iso8601?(value)

Test if a value is an ISO8601 encoded string.

is_snowflake(value)
(macro)

is_snowflake?(value)

Specs

is_snowflake?(any()) :: boolean()

Function to test if a term is an integer snowflake or binary snowflake

is_unixtime(value)
(macro)

is_unixtime?(value)

Specs

is_unixtime?(any()) :: boolean()

Check if the term is a valid DateTime.t/0.

to_datetime(value)

Specs

to_datetime(any()) :: DateTime.t() | nil | :error

Convert a value to an elixir DateTime

Examples

iex> Remedy.TimeHelpers.to_datetime(nil) nil

iex> Remedy.TimeHelpers.to_datetime(15824000000000) ~U[2015-01-01T01:02:52.735Z]

iex> Remedy.TimeHelpers.to_datetime(919970797920067654) ~U[2021-12-13T15:15:30.455Z]

iex> Remedy.TimeHelpers.to_datetime("2021-12-13T15:13:33.774426Z") ~U[2021-12-13T15:13:33.774426Z]

iex> Remedy.TimeHelpers.to_datetime(:some_atom) :error

to_iso8601(value)
(since 0.6.8)

Specs

to_iso8601(any()) :: Remedy.ISO8601.t() | nil | :error

Convert a value to a string representation of the ISO8601 format.

Examples

iex Remedy.TimeHelpers.to_iso8601(nil)
nil

iex> Remedy.TimeHelpers.to_iso8601(15824000000000)
"2015-01-01T01:02:52.735Z"

iex> Remedy.TimeHelpers.to_iso8601(919970797920067654)
"2021-12-13T15:15:30.455Z"

iex> Remedy.TimeHelpers.to_iso8601("2021-12-13T15:13:33.774426Z")
"2021-12-13T15:13:33.774426Z"

iex Remedy.TimeHelpers.to_iso8601(:some_atom)
:error

to_snowflake(value)

Specs

to_snowflake(any()) :: Remedy.Snowflake.t() | :error

to_unixtime(value)

Specs

to_unixtime(any()) :: Timestamp.t() | nil | :error

Convert a value to its unix time.

Examples

iex Remedy.TimeHelpers.to_unixtime(nil)
nil

iex> Remedy.TimeHelpers.to_unixtime(797634639408530455)
1610241317369

iex> Remedy.TimeHelpers.to_unixtime("2021-01-10T01:15:17.369Z")
1610241317369

iex> date_time = DateTime.now!("Etc/UTC")
...> Remedy.TimeHelpers.to_unixtime(date_time)

iex Remedy.TimeHelpers.to_unixtime(:some_atom)
:error