Remedy.Timestamp (Remedy v0.6.9)

Ecto.Type compatible Discord Timestamp type.

(The 3rd custom time format required for Discord.)

This is a standard Unix timestamp, with a millisecond precision.

Casting

The following are examples of valid inputs for casting. Regardless of the format provided, values will be cast to an integer/0 value for storage.

Snowflake

927056337051992064

DateTime

~U[2021-12-13T15:15:30.455Z]

ISO8601 String

"2020-01-01T00:00:00.000Z"

Unix

1577836800

Types

c()

Specs

c() :: t() | ISO8601.t() | DateTime.t() | integer() | String.t()

Castable to Discord Timestamp.

t()

Specs

t() :: 1_420_034_400_000..4_102_408_800_000

A Discord Timestamp.