Remedy.Type behaviour (Remedy v0.6.9)

Ecto.Type implementation of types.

Simply put, types are magic numbers for pre agreed values. Convenience does not lend itself to looking up the numbers when one of the fields needs to be used. The behaviour provides simplicity for these fields. Modules implementing this behaviour generally end with "Type". The complete list of Type types is:

Remedy.Schema.ActivityType

Remedy.Schema.AuditLogActionType

Remedy.Schema.AuditLogOptionType

Remedy.Schema.ButtonStyle

Remedy.Schema.CallbackType

Remedy.Schema.ChannelType

Remedy.Schema.CommandOptionType

Remedy.Schema.CommandPermissionType

Remedy.Schema.CommandType

Remedy.Schema.ComponentType

Remedy.Schema.EventEntityType

Remedy.Schema.EventPrivacyLevel

Remedy.Schema.EventStatus

Remedy.Schema.GuildDefaultMessageNotification

Remedy.Schema.GuildExplicitContentFilter

Remedy.Schema.GuildMFALevel

Remedy.Schema.GuildNSFWLevel

Remedy.Schema.GuildPremiumTier

Remedy.Schema.GuildVerificationLevel

Remedy.Schema.IntegrationExpireType

Remedy.Schema.IntegrationType

Remedy.Schema.MessageActivityType

Remedy.Schema.MessageType

Remedy.Schema.PermissionOverwriteType

Remedy.Schema.ResponseType

Remedy.Schema.StagePrivacyLevel

Remedy.Schema.StickerFormatType

Remedy.Schema.StickerType

Remedy.Schema.ThreadType

Remedy.Schema.WebhookType

Documentation for these types including how to cast them are available within the individual modules.

Callbacks

to_binary(any)

Specs

to_binary(any()) :: any()

to_integer(any)

Specs

to_integer(any()) :: any()