Remedy.CDN (Remedy v0.6.2)

Discord CDN interface.

Storing images and other assets can be painful, use this module to retreive assets from the discord cdn rather than storing them.

Each function takes the required parameters to directly access a resource.

Format

All images are returned as either a .png or a .gif if the asset is animated. This is done automatically.

Size

Images can be requested of a size in [16, 32, 64, 128, 256, 512, 1024, 2048, 4096]. This is given as an optional final argument to any of the functions in this module. Arguments given will be rounded to the next largest, or the largest size if you try to go over.

Missing Endpoints:

sticker

team icon

role icon

guild member icon

Types

discriminator()

Specs

discriminator() :: integer()

hash()

Specs

hash() :: binary()

size()

Specs

size() :: integer() | nil

snowflake()

Specs

snowflake() :: Snowflake.t()

Functions

application_cover(application, size \\ nil)

Specs

application_cover(Remedy.Schema.App.t(), size()) :: binary()

application_icon(application, size \\ nil)

Specs

application_icon(Remedy.Schema.App.t(), size()) :: binary()

custom_emoji(id, size \\ nil)

guild_banner(guild, size \\ nil)

Specs

guild_banner(Remedy.Schema.Guild.t(), size()) :: binary()

guild_discovery_splash(guild, size \\ nil)

Specs

guild_discovery_splash(Remedy.Schema.Guild.t(), size()) :: binary()

guild_icon(guild, size \\ nil)

Specs

guild_icon(Remedy.Schema.Guild.t(), size()) :: binary()

Returns the url for the guilds icon.

Examples

iex> Remedy.API.get_guild!(872417560094732328) |> Remedy.CDN.guild_icon()
"https://cdn.discordapp.com/icons/872417560094732328/f817c5adaf96672c94a17de8e944f427.jpg?size=128"

iex> Remedy.API.get_guild!(872417560094732328) |> Remedy.CDN.guild_icon(100)
"https://cdn.discordapp.com/icons/872417560094732328/f817c5adaf96672c94a17de8e944f427.jpg?size=128"

iex> Remedy.CDN.guild_icon(872417560094732328, "f817c5adaf96672c94a17de8e944f427", "cheese")
"https://cdn.discordapp.com/icons/872417560094732328/f817c5adaf96672c94a17de8e944f427.jpg"

guild_splash(guild, size \\ nil)

Specs

guild_splash(Remedy.Schema.Guild.t(), size()) :: binary()

user_avatar(user, size \\ nil)

Specs

user_avatar(Remedy.Schema.User.t(), size()) :: binary()

user_banner(user, size \\ nil)

Specs

user_banner(Remedy.Schema.User.t(), size()) :: binary()