Remedy.CDN (Remedy v0.6.9)

Discord CDN interface.

Format

All images are returned as either a .png or a .gif if the asset is animated. This is done automatically where possible.

Size

Images can be requested of a size in [16, 32, 64, 128, 256, 512, 1024, 2048, 4096]. This is given as an optional final argument to any of the functions in this module. Arguments given will be rounded to the next largest, or the largest size if you try to go over.

Unless specified, discord will deliver the smallest size available. ( Potato )

Types

discriminator()

Specs

discriminator() :: 0..9999

A four digit integer assigned by discord to avoid duplicate usernames

hash()

Specs

hash() :: binary()

The images hash.

size()

Specs

size() :: integer() | nil

The images size.

uri()

Specs

uri() :: String.t()

Functions

achievement_icon(application_id, achievement_id, icon_hash, size \\ nil)

Specs

achievement_icon(Remedy.Snowflake.t(), Remedy.Snowflake.t(), hash(), size()) ::
  uri()

Returns an achievement icon url.

Examples

iex> Remedy.CDN.achievement_icon(12345678901234567890123456789012, "f817c5adaf96672c94a17de8e944f427")
"https://cdn.discordapp.com/achievement-icons/12345678901234567890123456789012/f817c5adaf96672c94a17de8e944f427.png"

application_asset(application_id, asset_id, size \\ nil)

Specs

application_asset(Remedy.Snowflake.t(), hash(), size()) :: uri()

Returns an application asset url.

Examples

iex> Remedy.CDN.application_asset(12345678901234567890123456789012, "f817c5adaf96672c94a17de8e944f427")
"https://cdn.discordapp.com/app-assets/12345678901234567890123456789012/f817c5adaf96672c94a17de8e944f427.png"

application_cover(application_id, cover_image, size \\ nil)

Specs

application_cover(Remedy.Snowflake.t(), hash(), size()) :: uri()

Returns an application cover url.

Examples

iex> Remedy.CDN.application_cover(12345678901234567890123456789012, "f817c5adaf96672c94a17de8e944f427")
"https://cdn.discordapp.com/app-cover-images/12345678901234567890123456789012/f817c5adaf96672c94a17de8e944f427.png"

application_icon(application_id, icon, size \\ nil)

Specs

application_icon(Remedy.Snowflake.t(), hash(), size()) :: uri()

Returns the url for the applications icon.

Examples

iex> Remedy.CDN.application_icon(12345678901234567890123456789012, "f817c5adaf96672c94a17de8e944f427")
"https://cdn.discordapp.com/app-icons/12345678901234567890123456789012/f817c5adaf96672c94a17de8e944f427.png"

custom_emoji(id, size \\ nil)

Specs

custom_emoji(Remedy.Snowflake.t(), size()) :: binary()

Returns the url for a custom emoji.

Examples

iex> Remedy.CDN.emoji(12345678901234567890123456789012, 128)
"https://cdn.discordapp.com/emojis/12345678901234567890123456789012.png?size=128"

default_user_avatar(discriminator, size \\ nil)

Specs

default_user_avatar(discriminator(), size()) :: uri()

Returns the url for the default avatar of a user.

That is, their avatar if they do not have one set, based on their discriminator. It can also take just the discriminator if you wish to do so.

Examples

iex> Remedy.CDN.default_user_avatar(3)
"https://cdn.discordapp.com/embed/avatars/3.png"

guild_banner(guild_id, guild_banner, size)

Specs

guild_banner(Remedy.Snowflake.t(), hash(), size()) :: uri()

Returns the url for a guilds banner.

Examples

iex> Remedy.CDN.guild_banner(848619361782726696, "7ed6ea26b7a5e64f78ca5df202cf4d13")
"https://cdn.discordapp.com/banners/848619361782726696/"7ed6ea26b7a5e64f78ca5df202cf4d13"

guild_discovery_splash(guild_id, guild_discovery_splash, size \\ nil)

Specs

guild_discovery_splash(Remedy.Snowflake.t(), hash(), size()) :: uri()

Returns the url for a guilds discovery splash.

Examples

iex> Remedy.CDN.guild_discovery_splash(848619361782726696, "7ed6ea26b7a5e64f78ca5df202cf4d13")
"https://cdn.discordapp.com/splashes/848619361782726696/7ed6ea26b7a5e64f78ca5df202cf4d13.png"

guild_icon(id, guild_icon, size \\ nil)

Specs

guild_icon(Remedy.Snowflake.t(), hash(), size()) :: uri()

Returns the url for a guilds icon.

Examples

iex> Remedy.CDN.guild_icon(872417560094732328, "f817c5adaf96672c94a17de8e944f427")
"https://cdn.discordapp.com/icons/872417560094732328/f817c5adaf96672c94a17de8e944f427.png"

guild_splash(id, splash, size \\ nil)

Specs

guild_splash(Remedy.Snowflake.t(), hash(), size()) :: uri()

Returns the url for a guilds splash.

Examples

iex> Remedy.CDN.guild_splash(848619361782726696, "7ed6ea26b7a5e64f78ca5df202cf4d13")
"https://cdn.discordapp.com/splashes/848619361782726696/7ed6ea26b7a5e64f78ca5df202cf4d13.png"

sticker_pack_banner(banner_asset_id, size)

Specs

sticker_pack_banner(Remedy.Snowflake.t(), size()) :: uri()

Returns a sticker banner url.

user_avatar(id, user_avatar, size \\ nil)

Specs

user_avatar(Remedy.Snowflake.t(), hash(), size()) :: uri()

Returns the url for the users avatar.

Examples

iex> Remedy.CDN.user_avatar(707047919332884520, "1df05ae0f21a24c377e9a1051c2b6035", 128)
"https://cdn.discordapp.com/avatars/707047919332884520/1df05ae0f21a24c377e9a1051c2b6035.png?size=128"

user_banner(user_id, user_banner, size \\ nil)

Specs

user_banner(Remedy.Snowflake.t(), hash(), size()) :: uri()

Returns the url for a users banner.

Examples

iex> Remedy.CDN.user_banner(179255727561375744, "e625e858e48602248a69bcfdfa886ab4")
"https://cdn.discordapp.com/banners/179255727561375744/e625e858e48602248a69bcfdfa886ab4.png"