Remedy.Cache (Remedy v0.6.9)

Functions for interacting with the Cache.

The Cache is implemented in ETS.

Bang!

While undocumented to reduce clutter, all functions can be banged to return or raise.

Types

reason()

Specs

reason() :: String.t()

Functions

app()

bot()

drop_metadata(map)

Specs

drop_metadata(map()) :: any()

drop_unloaded_assoc(map)

Specs

drop_unloaded_assoc(map()) :: any()

fetch_ban(guild_id, user_id)

Specs

fetch_ban(Remedy.Snowflake.c(), Remedy.Snowflake.c()) ::
  {:error, reason()} | {:ok, Remedy.Schema.Ban.t()}

Fetch a ban from the Cache by user_id & guild_id

fetch_channel(id)

Specs

fetch_channel(Remedy.Snowflake.c()) ::
  {:error, reason()} | {:ok, Remedy.Schema.Channel.t()}

Fetch a channel from the cache.

fetch_emoji(id)

Fetch a emoji from the cache

fetch_guild(id)

Fetch a guild from the cache

fetch_role(role_id)

Specs

fetch_role(any()) :: {:error, :not_found} | {:ok, Remedy.Schema.Role.t()}

Fetch a role from the cache.

Returns {:ok, %Role{}} or {:error, reason}

fetch_user(user_id)

Fetch a user from the cache.

Returns {:ok, %User{}} or {:error, reason}

get_ban(guild_id, user_id)

Specs

get_ban(Remedy.Snowflake.c(), Remedy.Snowflake.c()) ::
  nil | Remedy.Schema.Ban.t()

get_member(guild_id, user_id)

list_bans()

Specs

list_bans() :: {:error, reason()} | {:ok, [Remedy.Schema.Ban.t()]}

List all bans

list_channels()

Specs

list_channels() :: {:error, term()} | {:ok, [Remedy.Schema.Channel.t()]}

list_channels(guild_id)

Specs

list_channels(Remedy.Snowflake.c()) ::
  {:error, term()} | {:ok, [Remedy.Schema.Channel.t()]}

List channels from the cache.

list_guild_bans(guild_id)

Specs

list_guild_bans(Remedy.Snowflake.c()) ::
  {:error, reason()} | {:ok, [Remedy.Schema.Ban.t()]}

List all bans associated with a guild.

list_guilds()

list_members()

List Members

list_user_bans(user_id)

Specs

list_user_bans(Remedy.Snowflake.c()) ::
  {:error, reason()} | {:ok, [Remedy.Schema.Ban.t()]}

List all bans associated with a user.

update_member(attrs)

user_banned?(guild_id, user_id)

Specs

user_banned?(Remedy.Snowflake.c(), Remedy.Snowflake.c()) :: true | false

Returns True/False to the user being banned from a guild.