Remedy.Consumer behaviour (Remedy v0.5.4)

Consumer process for gateway event handling.

Consuming Dispatch Events

To handle events, Remedy uses a GenStage implementation.

Remedy defines the producer and producer_consumer in the GenStage design. To consume the events you must create at least one consumer process.

Remedy uses a ConsumerSupervisor to dispatch events, meaning your handlers will each be ran in their own seperate task.

The full list of dispatch events and their inner payload is described in the type specs within this module.

Regular payloads are delivered after casting to their schema and ratified against the cache.

Irregular payloads are classified as those which do not directly map to a standard discord object. They undergo extensive manipulation prior to updating the cache. They are described under the DISPATCH section of the documentation.

Example

It is recommended that you supervise your consumers. First we set up a supervisor module for our consumers.

#example_supervisor.ex
defmodule MyApp.ExampleSupervisor do
  use Supervisor

  def start_link(args) do
    Supervisor.start_link(__MODULE__, args, name: __MODULE__)
  end

  @impl true
  def init(_init_arg) do
    children = [ExampleConsumer]

    Supervisor.init(children, strategy: :one_for_one)
  end
end

You can then set up your consumer module.

#example_consumer.ex
defmodule ExampleConsumer do
  use Remedy.Consumer

  alias Remedy.Api

  def start_link do
    Consumer.start_link(__MODULE__)
  end

  def handle_event({:MESSAGE_CREATE, %Remedy.Schema.Message{content: content}, _ws_state}) do
    case content do
      "!sleep" ->
        Api.create_message(msg.channel_id, "Going to sleep...")
        Process.sleep(3000)

      "!ping" ->
        Api.create_message(msg.channel_id, "pyongyang!")

      "!raise" ->
        raise "No problems here!"
    end
  end

  def handle_event(_event) do
    :noop
  end
end

Types

event()

Specs

event() ::
  {:CHANNEL_CREATE, Remedy.Schema.Channel.t(), Remedy.Gateway.WSState.t()}
  | {:CHANNEL_DELETE, Remedy.Schema.Channel.t(), Remedy.Gateway.WSState.t()}
  | {:CHANNEL_PINS_UPDATE, ChannelPinsUpdate.t(), Remedy.Gateway.WSState.t()}
  | {:CHANNEL_UPDATE, Remedy.Schema.Channel.t(), Remedy.Gateway.WSState.t()}
  | {:GUILD_AVAILABLE, Remedy.Schema.Guild.t(), Remedy.Gateway.WSState.t()}
  | {:GUILD_BAN_ADD, GuildBanAdd.t(), Remedy.Gateway.WSState.t()}
  | {:GUILD_BAN_REMOVE, GuildBanRemove.t(), Remedy.Gateway.WSState.t()}
  | {:GUILD_CREATE, Remedy.Schema.Guild.t(), Remedy.Gateway.WSState.t()}
  | {:GUILD_DELETE, Remedy.Schema.Guild.t(), Remedy.Gateway.WSState.t()}
  | {:GUILD_EMOJIS_UPDATE, [Remedy.Schema.Emoji.t()],
     Remedy.Gateway.WSState.t()}
  | {:GUILD_INTEGRATIONS_UPDATE, GuildIntegrationsUpdate.t(),
     Remedy.Gateway.WSState.t()}
  | {:GUILD_MEMBER_ADD, Remedy.Schema.Member.t(), Remedy.Gateway.WSState.t()}
  | {:GUILD_MEMBER_REMOVE, Remedy.Schema.Member.t(), Remedy.Gateway.WSState.t()}
  | {:GUILD_MEMBER_UPDATE, Remedy.Schema.Member.t(), Remedy.Schema.Member.t(),
     Remedy.Gateway.WSState.t()}
  | {:GUILD_MEMBERS_CHUNK, map(), Remedy.Gateway.WSState.t()}
  | {:GUILD_ROLE_CREATE, Remedy.Schema.Role.t(), Remedy.Gateway.WSState.t()}
  | {:GUILD_ROLE_DELETE, Remedy.Schema.Role.t(), Remedy.Gateway.WSState.t()}
  | {:GUILD_ROLE_UPDATE, Remedy.Schema.Role.t(), Remedy.Gateway.WSState.t()}
  | {:GUILD_UNAVAILABLE, UnavailableGuild.t(), Remedy.Gateway.WSState.t()}
  | {:GUILD_UPDATE, Remedy.Schema.Guild.t(), Remedy.Gateway.WSState.t()}
  | {:MESSAGE_ACK, map(), Remedy.Gateway.WSState.t()}
  | {:MESSAGE_CREATE, Remedy.Schema.Message.t(), Remedy.Gateway.WSState.t()}
  | {:MESSAGE_DELETE_BULK, MessageDeleteBulk.t(), Remedy.Gateway.WSState.t()}
  | {:MESSAGE_DELETE, MessageDelete.t(), Remedy.Gateway.WSState.t()}
  | {:MESSAGE_REACTION_ADD, MessageReactionAdd.t(), Remedy.Gateway.WSState.t()}
  | {:MESSAGE_REACTION_REMOVE_ALL, MessageReactionRemoveAll.t(),
     Remedy.Gateway.WSState.t()}
  | {:MESSAGE_REACTION_REMOVE_EMOJI, MessageReactionRemoveEmoji.t(),
     Remedy.Gateway.WSState.t()}
  | {:MESSAGE_REACTION_REMOVE, MessageReactionRemove.t(),
     Remedy.Gateway.WSState.t()}
  | {:MESSAGE_UPDATE, Remedy.Schema.Message.t(), Remedy.Gateway.WSState.t()}
  | {:PRESENCE_UPDATE, Presence.t(), Remedy.Gateway.WSState.t()}
  | {:READY, Ready.t(), Remedy.Gateway.WSState.t()}
  | {:RESUMED, map(), Remedy.Gateway.WSState.t()}
  | {:TYPING_START, TypingStart.t(), Remedy.Gateway.WSState.t()}
  | {:USER_UPDATE, Remedy.Schema.User.t(), Remedy.Gateway.WSState.t()}
  | {:VOICE_READY, VoiceReady.t(), VoiceWSState.t()}
  | {:VOICE_SERVER_UPDATE, VoiceServerUpdate.t(), Remedy.Gateway.WSState.t()}
  | {:VOICE_SPEAKING_UPDATE, SpeakingUpdate.t(), VoiceWSState.t()}
  | {:VOICE_STATE_UPDATE, VoiceState.t(), Remedy.Gateway.WSState.t()}
  | {:WEBHOOKS_UPDATE, map(), Remedy.Gateway.WSState.t()}

Callbacks

handle_event(event)

Specs

handle_event(event()) :: any()

handle_event(any()) :: :noop