Remedy.Consumer behaviour (Remedy v0.6.9)

Consumer process for gateway event handling.

You have two options for handling events. You can create a consumer module through the behaviour described in this module, which implements handle_event/1. Or you can create your own GenStage or Broadway pipeline to process the events. Details of each of these methods are described below. Events will be buffered up to 10,000 inside the Remedy.Consumer.Producer process waiting for a subscriber to connect. If you are using the Consumer Module method, you will subscribe immediately. If you are implementing your own GenStage or Broadway pipeline, you will have to start your module and subscribe yourself.

Consumer Module

The simplest consumer implementation is to use the Consumer behaviour by adding use Remedy.Consumer to your module.

This will handle all of the supervision and set up for you. all you need to implement is the handle_event/1 callback for all the events you wish to consume. Using this method will start your consumer under the Remedy supervision tree and you are not required to start it yourself.

This method is used to implement the standard event logging module, which looks something like this internally.

defmodule Remedy.Consumer.LumberJack do
  use Remedy.Consumer

  def handle_event(event, payload, _meta) do
    Logger.debug("#{inspect(event)} #{inspect(payload)}")
  end

end

Internally, the events are run in a ConsumerSupervisor which is configured to run System.schedulers/0 number of tasks concurrently. The tasks are wrapped in a try rescue block to ensure the consumer will never be taken down. If you start seeing CONSUMER_ERROR in your logs it means there is an error with your callback implementation and your callbacks are being rescued. There is no retry logic while using this method and the event is lost.

Genstage

If you wish to implement a custom GenStage pipeline you can simply subscribe to Remedy.Consumer.Producer inside your init callback. The other implementation details are the concern of the readers individual situation, such as how the module will be started and supervised.

defmodule MyApp.PipelineProducer do
  @moduledoc false
  use GenStage

  @impl GenStage
  def init(_opts) do
    opts = [
      dispatcher: GenStage.DemandDispatcher,
      subscribe_to: [Remedy.Consumer.Producer]
    ]
    {:producer_consumer, :state, opts}
  end

  @impl GenStage
  def handle_events(events, _from, state) do
    {:noreply, [events], state, :hibernate}
  end

  @impl GenStage
  def handle_demand(_incoming_demand, state) do
    {:noreply, [], state, :hibernate}
  end
end

Broadway

To implement a Broadway pipeline you should first create a :producer_consumer per the example above. You can then use that producer as your as your Broadway producer. Following that you just need to implement the callbacks specific to your pipeline. Note that because Remedy uses Broadway internally, there is no need to include it as a dependency.

defmodule MyApp.Pipeline do
  use Broadway
  def start_link(_opts) do
    Broadway.start_link(__MODULE__,
      name: __MODULE__,
      producer: [
        module: {PipelineProducer, []},
        concurrency: 1
      ],
      processors: [
        default: [concurrency: 10]
      ],
      batchers: [
        channel_create: [concurrency: 100],
        default: [concurrency: 10],
        unhandled: [concurrency: 1]
      ]
    )
  end

  def handle_message(_, %{data: %{event: :CHANNEL_CREATE}}} = message, _) do
    message
    |> Broadway.Message.put_batcher(:CHANNEL_CREATE)
  end

  # Other Callbacks
end

Types

channel_create()

Specs

channel_create() ::
  {:CHANNEL_CREATE, Remedy.Schema.Channel.t(), Remedy.Consumer.Metadata.t()}

Received when a channel is created.

channel_delete()

Specs

channel_delete() ::
  {:CHANNEL_DELETE, Remedy.Schema.Channel.t(), Remedy.Consumer.Metadata.t()}

Received when a channel is deleted.

channel_pins_update()

Specs

channel_pins_update() ::
  {:CHANNEL_UPDATE, Remedy.Schema.Channel.t(), Remedy.Consumer.Metadata.t()}

Received when a message is pinned or unpinned.

channel_update()

Specs

channel_update() ::
  {:CHANNEL_UPDATE, Remedy.Schema.Channel.t(), Remedy.Consumer.Metadata.t()}

Received when a channel is updated.

guild_available()

Specs

guild_available() ::
  {:GUILD_AVAILABLE, Remedy.Schema.Guild.t(), Remedy.Consumer.Metadata.t()}

guild_ban_add()

Specs

guild_ban_add() ::
  {:GUILD_BAN_ADD, Remedy.Schema.Ban.t(), Remedy.Consumer.Metadata.t()}

Received when a user is banned from a guild.

guild_ban_remove()

Specs

guild_ban_remove() ::
  {:GUILD_BAN_REMOVE, Remedy.Schema.Ban.t(), Remedy.Consumer.Metadata.t()}

Received when a user is unbanned from a guild.

guild_create()

Specs

guild_create() ::
  {:GUILD_CREATE, Remedy.Schema.Guild.t(), Remedy.Consumer.Metadata.t()}

Received when joining a new guild.

guild_delete()

Specs

guild_delete() ::
  {:GUILD_DELETE, Remedy.Schema.Guild.t(), Remedy.Consumer.Metadata.t()}

guild_emojis_update()

Specs

guild_emojis_update() ::
  {:GUILD_EMOJIS_UPDATE, Remedy.Schema.Guild.t(), Remedy.Consumer.Metadata.t()}

Received when a guild's emojis have been updated.

guild_integrations_update()

Specs

guild_integrations_update() ::
  {:GUILD_INTEGRATIONS_UPDATE, Remedy.Schema.Guild.t(),
   Remedy.Consumer.Metadata.t()}

Received when a guild integration is updated.

guild_kicked()

Specs

guild_kicked() ::
  {:GUILD_KICKED, Remedy.Schema.Guild.t(), Remedy.Consumer.Metadata.t()}

Recieved when the bot is kicked from a guild.

guild_leave()

Specs

guild_leave() ::
  {:GUILD_LEAVE, Remedy.Schema.Guild.t(), Remedy.Consumer.Metadata.t()}

Received when the bot leaves a guild.

guild_member_add()

Specs

guild_member_add() ::
  {:GUILD_MEMBER_ADD, Remedy.Schema.Member.t(), Remedy.Consumer.Metadata.t()}

Received when a new user joins a guild.

guild_member_kicked()

Specs

guild_member_kicked() ::
  {:GUILD_MEMBER_REMOVE, Remedy.Schema.Member.t(), Remedy.Consumer.Metadata.t()}

Received when a used is removed from a guild.

guild_member_update()

Specs

guild_member_update() ::
  {:GUILD_MEMBER_UPDATE, Remedy.Schema.Member.t(), Remedy.Consumer.Metadata.t()}

Received when a guild member is updated.

guild_role_create()

Specs

guild_role_create() ::
  {:GUILD_ROLE_CREATE, Remedy.Schema.Role.t(), Remedy.Consumer.Metadata.t()}

Received when a guild role is created.

guild_role_delete()

Specs

guild_role_delete() ::
  {:GUILD_ROLE_DELETE, Remedy.Schema.Role.t(), Remedy.Consumer.Metadata.t()}

Received when a guild role is deleted.

guild_role_update()

Specs

guild_role_update() ::
  {:GUILD_ROLE_UPDATE, Remedy.Schema.Role.t(), Remedy.Consumer.Metadata.t()}

Received when a guild role is updated.

guild_unavailable()

Specs

guild_unavailable() ::
  {:GUILD_UNAVAILABLE, Remedy.Schema.Guild.t(), Remedy.Consumer.Metadata.t()}

Received when a guild is unavailable due to an outage.

guild_update()

Specs

guild_update() ::
  {:GUILD_UPDATE, Remedy.Schema.Guild.t(), Remedy.Consumer.Metadata.t()}

Receivedent when a guild is updated.

integration_create()

Specs

integration_create() ::
  {:INTEGRATION_CREATE, Remedy.Schema.Integration.t(),
   Remedy.Consumer.Metadata.t()}

Received when an integration is created.

integration_delete()

Specs

integration_delete() ::
  {:INTEGRATION_DELETE, Remedy.Schema.Integration.t(),
   Remedy.Consumer.Metadata.t()}

Received when an integration is deleted.

integration_update()

Specs

integration_update() ::
  {:INTEGRATION_UPDATE, Remedy.Schema.Integration.t(),
   Remedy.Consumer.Metadata.t()}

Received when an integration is updated.

interaction_create()

Specs

interaction_create() ::
  {:INTERACTION_CREATE, Remedy.Schema.Interaction.t(),
   Remedy.Consumer.Metadata.t()}

Received when a user triggers an Application Command

message_create()

Specs

message_create() ::
  {:MESSAGE_CREATE, Remedy.Schema.Message.t(), Remedy.Consumer.Metadata.t()}

Received when a message is created.

message_delete()

Specs

message_delete() ::
  {:MESSAGE_DELETE, Remedy.Schema.Message.t(), Remedy.Consumer.Metadata.t()}

Received when a messgae is deleted.

message_delete_bulk()

Specs

message_delete_bulk() ::
  {:MESSAGE_DELETE_BULK, MessageDeleteBulk.t(), Remedy.Consumer.Metadata.t()}

Received when multiple messages are deleted at once.

message_update()

Specs

message_update() ::
  {:MESSAGE_UPDATE, Remedy.Schema.Message.t(), Remedy.Consumer.Metadata.t()}

Sent when a message is updated.

presence_update()

Specs

presence_update() ::
  {:PRESENCE_UPDATE, Remedy.Schema.User.t(), Remedy.Consumer.Metadata.t()}

Received when a user's presence or info is updated.

reaction_add()

Specs

reaction_add() ::
  {:REACTION_ADD, Remedy.Schema.Reaction.t(), Remedy.Consumer.Metadata.t()}

Received when a user adds a reaction to a message.

reaction_remove()

Specs

reaction_remove() ::
  {:REACTION_REMOVE, Remedy.Schema.Reaction.t(), Remedy.Consumer.Metadata.t()}

Received when a user removes a reaction from a message.

ready()

Specs

ready() :: {:READY, Ready.t(), Remedy.Consumer.Metadata.t()}

Received when the bot is online.

thread_create()

Specs

thread_create() ::
  {:THREAD_CREATE, Remedy.Schema.Thread.t(), Remedy.Consumer.Metadata.t()}

Sent when a thread is created or when the user is added to a thread.

thread_delete()

Specs

thread_delete() ::
  {:THREAD_DELETE, Remedy.Schema.Thread.t(), Remedy.Consumer.Metadata.t()}

Sent when a thread relevant to the current user is deleted.

thread_member_update()

Specs

thread_member_update() ::
  {:THREAD_MEMBER_UPDATE, Remedy.Schema.ThreadMember.t(),
   Remedy.Consumer.Metadata.t()}

Received when the thread member object for the current user is updated.

thread_members_update()

Specs

thread_members_update() ::
  {:THREAD_MEMBERS_UPDATE, Remedy.Schema.ThreadMember.t(),
   Remedy.Consumer.Metadata.t()}

Received when anyone is added to or removed from a thread.

thread_update()

Specs

thread_update() ::
  {:THREAD_UPDATE, Remedy.Schema.Thread.t(), Remedy.Consumer.Metadata.t()}

Received when a thread is updated.

typing_start()

Specs

typing_start() :: {:TYPING_START, TypingStart.t(), Remedy.Consumer.Metadata.t()}

Received when a user begins typing in a channel.

user_update()

Specs

user_update() ::
  {:USER_UPDATE, Remedy.Schema.User.t(), Remedy.Consumer.Metadata.t()}

Received when a user is updated.

voice_ready()

Specs

voice_ready() :: {:VOICE_READY, VoiceReady.t(), Remedy.Consumer.Metadata.t()}

Received when a webhook is updated.

voice_server_update()

Specs

voice_server_update() ::
  {:VOICE_SERVER_UPDATE, VoiceServerUpdate.t(), Remedy.Consumer.Metadata.t()}

Received when a webhook is updated.

voice_speaking_update()

Specs

voice_speaking_update() ::
  {:VOICE_SPEAKING_UPDATE, SpeakingUpdate.t(), Remedy.Consumer.Metadata.t()}

Received when a webhook is updated.

voice_state_update()

Specs

voice_state_update() ::
  {:VOICE_STATE_UPDATE, Remedy.Schema.VoiceState.t(),
   Remedy.Consumer.Metadata.t()}

Received when a user's voice state is updated.

webhooks_update()

Specs

webhooks_update() ::
  {:WEBHOOKS_UPDATE, WebhooksUpdate.t(), Remedy.Consumer.Metadata.t()}

Received when a webhook is updated.

Callbacks

handle_event(term)
(optional)

Specs

handle_event(term()) :: term() | :ok | :noop