Remedy.ImageData (Remedy v0.6.9)

Ecto.Type implementation of Image Data.

This allows a URL or path to be provided and the image data will be constructed from the linked image.

This is only used with certain API endpoints and should not be used as a general purpose type for storing images in Ecto.

Casting

The following are examples of valid inputs for casting. Regardless of the format provided, values will be cast to an binary/0 value for storage.

Image Data

"data:image/jpeg;base64,BASE64_ENCODED_JPEG_IMAGE_DATA"

Image URL

"https://www.google.com/images/branding/googlelogo/1x/googlelogo_color_272x92dp.png"

Note: Using this type with a URL will use external connections outside of the http worker pool. See the documentation on Connection limits for considerations.

Types

c()

Specs

c() :: Path.t() | URI.t() | String.t()

Castable to Image.

t()

Specs

t() :: 0..16_777_215

A so called Image Type.

Functions

info(_)