Remedy.Schema.Permission (Remedy v0.6.9)

Ecto.Type compatible bit flag type.

Bit	Flag

1 << 0	:CREATE_INSTANT_INVITE

1 << 1	:KICK_MEMBERS

1 << 2	:BAN_MEMBERS

1 << 3	:ADMINISTRATOR

1 << 4	:MANAGE_CHANNELS

1 << 5	:MANAGE_GUILD

1 << 6	:ADD_REACTIONS

1 << 7	:VIEW_AUDIT_LOG

1 << 8	:PRIORITY_SPEAKER

1 << 9	:STREAM

1 << 10	:VIEW_CHANNEL

1 << 11	:SEND_MESSAGES

1 << 12	:SEND_TTS_MESSAGES

1 << 13	:MANAGE_MESSAGES

1 << 14	:EMBED_LINKS

1 << 15	:ATTACH_FILES

1 << 16	:READ_MESSAGE_HISTORY

1 << 17	:MENTION_EVERYONE

1 << 18	:USE_EXTERNAL_EMOJIS

1 << 19	:VIEW_GUILD_INSIGHTS

1 << 20	:CONNECT

1 << 21	:SPEAK

1 << 22	:MUTE_MEMBERS

1 << 23	:DEAFEN_MEMBERS

1 << 24	:MOVE_MEMBERS

1 << 25	:USE_VAD

1 << 26	:CHANGE_NICKNAME

1 << 27	:MANAGE_NICKNAMES

1 << 28	:MANAGE_ROLES

1 << 29	:MANAGE_WEBHOOKS

1 << 30	:MANAGE_EMOJIS_AND_STICKERS

1 << 31	:USE_APPLICATION_COMMANDS

1 << 32	:REQUEST_TO_SPEAK

1 << 33	:MANAGE_EVENTS

1 << 34	:MANAGE_THREADS

1 << 35	:USE_PUBLIC_THREADS

1 << 36	:USE_PRIVATE_THREADS

1 << 37	:USE_EXTERNAL_STICKERS

1 << 38	:SEND_MESSAGES_IN_THREADS

1 << 39	:START_EMBEDDED_ACTIVITIES

1 << 40	:MODERATE_MEMBERS

Casting

The following are examples of valid inputs for casting. Regardless of the format provided, values will be cast to an integer/0 value for storage.

Decimal Integer

2147484160

Binary Integer

0b100000000000000000000000000000

String List

["MOVE_MEMBERS", "READ_MESSAGE_HISTORY", "STREAM"]

Atom List

[:MANAGE_MESSAGES, :MODERATE_MEMBERS, :READ_MESSAGE_HISTORY,
 :START_EMBEDDED_ACTIVITIES]

Boolean Map

%{
  ATTACH_FILES: true,
  MANAGE_EMOJIS_AND_STICKERS: true,
  MENTION_EVERYONE: true,
  MODERATE_MEMBERS: true,
  PRIORITY_SPEAKER: true,
  SEND_MESSAGES_IN_THREADS: true,
  SPEAK: true,
  USE_EXTERNAL_EMOJIS: true,
  USE_PUBLIC_THREADS: true,
  VIEW_AUDIT_LOG: true
}

Types

c()

Specs

c() :: t() | [f()] | map() | [String.t()]

Castable to Permission Type

f()

Specs

f() ::
  :START_EMBEDDED_ACTIVITIES
  | :VIEW_AUDIT_LOG
  | :CREATE_INSTANT_INVITE
  | :CHANGE_NICKNAME
  | :USE_APPLICATION_COMMANDS
  | :VIEW_CHANNEL
  | :MENTION_EVERYONE
  | :DEAFEN_MEMBERS
  | :SPEAK
  | :USE_EXTERNAL_EMOJIS
  | :EMBED_LINKS
  | :MOVE_MEMBERS
  | :READ_MESSAGE_HISTORY
  | :MANAGE_NICKNAMES
  | :VIEW_GUILD_INSIGHTS
  | :MUTE_MEMBERS
  | :ATTACH_FILES
  | :BAN_MEMBERS
  | :USE_VAD
  | :MANAGE_GUILD
  | :MANAGE_WEBHOOKS
  | :MANAGE_ROLES
  | :USE_PUBLIC_THREADS
  | :MODERATE_MEMBERS
  | :MANAGE_EVENTS
  | :SEND_TTS_MESSAGES
  | :STREAM
  | :USE_EXTERNAL_STICKERS
  | :MANAGE_EMOJIS_AND_STICKERS
  | :CONNECT
  | :SEND_MESSAGES
  | :MANAGE_MESSAGES
  | :USE_PRIVATE_THREADS
  | :REQUEST_TO_SPEAK
  | :ADD_REACTIONS
  | :PRIORITY_SPEAKER
  | :MANAGE_CHANNELS
  | :SEND_MESSAGES_IN_THREADS
  | :KICK_MEMBERS
  | :ADMINISTRATOR
  | :MANAGE_THREADS

Permission Flags

load()

Specs

load() :: [f()]

Loaded Permission Flags

t()

Specs

t() :: 0..2_199_023_255_551

Permission Type