Remedy.Schema.UserFlags (Remedy v0.6.9)

User Flags Type

Bit	Flag

1 << 0	:DISCORD_EMPLOYEE

1 << 1	:PARTNERED_SERVER_OWNER

1 << 2	:HYPESQUAD_EVENTS

1 << 3	:BUG_HUNTER_LEVEL_1

1 << 6	:HYPESQUAD_BRAVERY

1 << 7	:HYPESQUAD_BRILLIANCE

1 << 8	:HYPESQUAD_BALANCE

1 << 9	:EARLY_NITRO_SUPPORTER

1 << 10	:TEAM_PSEUDO_USER

1 << 12	:SYSTEM

1 << 14	:BUG_HUNTER_LEVEL_2

1 << 16	:VERIFIED_BOT

1 << 17	:VERIFIED_DEVELOPER

1 << 18	:CERTIFIED_MODERATOR

1 << 19	:BOT_HTTP_INTERACTIONS

Representations

The following representations are all valid when casting to a schema. However the underlying data will always be stored as an integer, and displayed as a list of strings.

Decimal Integer

1005519

Binary Integer

0b11110101011111001111

String List

["BOT_HTTP_INTERACTIONS", "BUG_HUNTER_LEVEL_1", "BUG_HUNTER_LEVEL_2",
 "CERTIFIED_MODERATOR", "DISCORD_EMPLOYEE", "EARLY_NITRO_SUPPORTER",
 "HYPESQUAD_BALANCE", "HYPESQUAD_BRAVERY", "HYPESQUAD_BRILLIANCE",
 "HYPESQUAD_EVENTS", "PARTNERED_SERVER_OWNER", "SYSTEM", "TEAM_PSEUDO_USER",
 "VERIFIED_BOT", "VERIFIED_DEVELOPER"]

Boolean Map

%{
  BOT_HTTP_INTERACTIONS: true,
  BUG_HUNTER_LEVEL_1: true,
  BUG_HUNTER_LEVEL_2: true,
  CERTIFIED_MODERATOR: true,
  DISCORD_EMPLOYEE: true,
  EARLY_NITRO_SUPPORTER: true,
  HYPESQUAD_BALANCE: true,
  HYPESQUAD_BRAVERY: true,
  HYPESQUAD_BRILLIANCE: true,
  HYPESQUAD_EVENTS: true,
  PARTNERED_SERVER_OWNER: true,
  SYSTEM: true,
  TEAM_PSEUDO_USER: true,
  VERIFIED_BOT: true,
  VERIFIED_DEVELOPER: true
}

Types

t()

Specs

t() :: 0..1_005_519